Crate ethcore

Ethcore library

Rust version:

nightly

Supported platforms:

OSX

Linux

Building:

Ubuntu 14.04 and later:

# install rustup
curl https://sh.rustup.rs -sSf | sh

# download and build parity
git clone https://github.com/paritytech/parity
cd parity
cargo build --release

OSX:

# install rocksdb && rustup
brew update
curl https://sh.rustup.rs -sSf | sh

# download and build parity
git clone https://github.com/paritytech/parity
cd parity
cargo build --release

Re-exports

pub extern crate ethstore;

pub use types::*;

pub use executive::contract_address;

Modules

account_provider
Account management.

block
Base data structure of this module is Block.

client
Blockchain database client.

db
Database utilities and definitions.

encoded
Lazily-decoded owning views of RLP-encoded blockchain objects. These views are meant to contain trusted data -- without encoding errors or inconsistencies.

engines
Consensus engine specification and basic implementations.

error
General error types for use in ethcore.

ethereum
Ethereum protocol module.

executed
Transaction execution format module.

executive
Transaction Execution environment.

header
Block header.

machine
Ethereum-like state machine definition.

miner
Miner module Keeps track of transactions and currently sealed pending block.

pod_state
State of all accounts in the system expressed in Plain Old Data.

snapshot
Snapshot creation, restoration, and network service.

spec
Blockchain params.

state
A mutable state representation suitable to execute transactions. Generic over a Backend. Deals with Accounts. Unconfirmed sub-states are managed with checkpoints which may be canonicalized or rolled back.

state_db
State database abstraction. For more info, see the doc for StateDB

test_helpers
Set of different helpers for client tests

trace
Tracing

verification
Block verification utilities.

views
Block oriented views onto rlp.

Macros

view

Enums

CreateContractAddress
Specifies how an address is calculated for a new contract.

Traits

BlockChainDB
Database backing BlockChain.

BlockChainDBHandler
Generic database handler. This trait contains one function open. When called, it opens database with a predefined config.