Struct ethcore::account::Account

pub struct Account {
    // some fields omitted
}

Single account in the system.

Methods

impl Account

fn new(balance: U256, nonce: U256, storage: HashMap<H256, H256>, code: Bytes) -> Account

General constructor.

fn from_pod(pod: PodAccount) -> Account

General constructor.

fn new_basic(balance: U256, nonce: U256) -> Account

Create a new account with the given balance.

fn from_rlp(rlp: &[u8]) -> Account

Create a new account from RLP.

fn new_contract(balance: U256) -> Account

Create a new contract account. NOTE: make sure you use init_code on this before commiting.

fn reset_code(&mut self)

Reset this account to the status of a not-yet-initialised contract. NOTE: Account should have init_code() called on it later.

fn init_code(&mut self, code: Bytes)

Set this account's code to the given code. NOTE: Account should have been created with new_contract() or have reset_code() called on it.

fn set_storage(&mut self, key: H256, value: H256)

Set (and cache) the contents of the trie's storage at key to value.

fn storage_at(&self, db: &HashDB, key: &H256) -> H256

Get (and cache) the contents of the trie's storage at key.

fn balance(&self) -> &U256

return the balance associated with this account.

fn nonce(&self) -> &U256

return the nonce associated with this account.

fn code_hash(&self) -> H256

return the code hash associated with this account.

fn code(&self) -> Option<&[u8]>

returns the account's code. If None then the code cache isn't available - get someone who knows to call note_code.

fn note_code(&mut self, code: Bytes) -> Result<(), H256>

Provide a byte array which hashes to the code_hash. returns the hash as a result.

fn is_cached(&self) -> bool

Is code_cache valid; such that code is going to return Some?

fn cache_code(&mut self, db: &HashDB) -> bool

Provide a database to lookup code_hash. Should not be called if it is a contract without code.

fn base_root(&self) -> &H256

return the storage root associated with this account.

fn storage_is_clean(&self) -> bool

Determine whether there are any un-commit()-ed storage-setting operations.

fn storage_root(&self) -> Option<&H256>

return the storage root associated with this account or None if it has been altered via the overlay.

fn recent_storage_root(&self) -> &H256

return the storage root associated with this account or None if it has been altered via the overlay.

fn storage_overlay(&self) -> Ref<HashMap<H256, (Filth, H256)>>

return the storage overlay.

fn inc_nonce(&mut self)

Increment the nonce of the account by one.

fn add_balance(&mut self, x: &U256)

Increment the nonce of the account by one.

fn sub_balance(&mut self, x: &U256)

Increment the nonce of the account by one.

fn commit_storage(&mut self, db: &mut HashDB)

Commit the storage_overlay to the backing DB and update storage_root.

fn commit_code(&mut self, db: &mut HashDB)

Commit any unsaved code. code_hash will always return the hash of the code_cache after this.

fn rlp(&self) -> Bytes

Export to RLP.