Module ethcore::account_provider

Account management.

Structs

AccountProvider
Account management. Responsible for unlocking accounts.

AccountProviderSettings
Account management settings.

DappId
Dapp identifier

HardwareWalletManager
Hardware wallet management interface.

IndexDerivation
Node in hierarchical derivation.

KeyFile
Public opaque type representing serializable KeyFile.

Signature
Signature encoded as RSV components

Transaction
A set of information describing an externally-originating message call or contract creation operation.

TransactionInfo
This is the transaction info we need to supply to Trezor message. It's more or less a duplicate of ethcore::transaction::Transaction, but we can't import ethcore here as that would be a circular dependency.

Enums

Action
Transaction action type.

Derivation
Derivation scheme for keys

HardwareError
Hardware wallet error.

KeyPath
Key derivation paths used on hardware wallets.

SignError
Signing error

Type Definitions

Error
AccountProvider errors.