Struct ethcore::account_provider::HardwareWalletManager

pub struct HardwareWalletManager { /* fields omitted */ }

Hardware wallet management interface.

Methods

impl HardwareWalletManager

pub fn new() -> Result<HardwareWalletManager, Error>

Hardware wallet constructor

pub fn set_key_path(&self, key_path: KeyPath)

Select key derivation path for a chain. Currently, only one hard-coded keypath is supported It is managed by ethcore/account_provider

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn list_wallets(&self) -> Vec<WalletInfo>

List connected wallets. This only returns wallets that are ready to be used.

pub fn list_locked_wallets(&self) -> Result<Vec<String>, Error>

Return a list of paths to locked hardware wallets This is only applicable to Trezor because Ledger only appears as a device when it is unlocked

pub fn wallet_info(&self, address: &H160) -> Option<WalletInfo>

Get connected wallet info.

pub fn sign_message(
&self,
address: &H160,
msg: &[u8]
) -> Result<Signature, Error>

Sign a message with the wallet (only supported by Ledger)

pub fn sign_transaction(
&self,
address: &H160,
t_info: &TransactionInfo,
encoded_transaction: &[u8]
) -> Result<Signature, Error>

Sign transaction data with wallet managing address.

pub fn pin_matrix_ack(&self, path: &str, pin: &str) -> Result<bool, Error>

Send a pin to a device at a certain path to unlock it This is only applicable to Trezor because Ledger only appears as a device when it is unlocked