Module ethcore::block

Blockchain block.

Structs

Block
A block, encoded as it is on the block chain.

ClosedBlock
Just like OpenBlock, except that we've applied Engine::on_close_block, finished up the non-seal header fields, and collected the uncles.

ExecutedBlock
An internal type for a block's common elements.

LockedBlock
Just like ClosedBlock except that we can't reopen it and it's faster.

OpenBlock
Block that is ready for transactions to be added.

SealedBlock
A block that has a valid seal.

Traits

Drain
Trait for a object that has a state database.

IsBlock
Trait for a object that is a ExecutedBlock.

Functions

enact_verified
Enact the block given by block_bytes using engine on the database db with given parent block header