Struct ethcore::block::Block

pub struct Block {
    pub header: Header,
    pub transactions: Vec<UnverifiedTransaction>,
    pub uncles: Vec<Header>,
}

A block, encoded as it is on the block chain.

Fields
header: Header

The header of this block.

transactions: Vec<UnverifiedTransaction>

The transactions in this block.

uncles: Vec<Header>

The uncles of this block.

Methods

impl Block

pub fn is_good(b: &[u8]) -> bool

Returns true if the given bytes form a valid encoding of a block in RLP.

pub fn rlp_bytes(&self) -> Bytes

Get the RLP-encoding of the block with the seal.