Struct ethcore::block::ClosedBlock

pub struct ClosedBlock { /* fields omitted */ }

Just like OpenBlock, except that we've applied Engine::on_close_block, finished up the non-seal header fields, and collected the uncles.

There is no function available to push a transaction.

Methods

impl ClosedBlock

pub fn hash(&self) -> H256

Get the hash of the header without seal arguments.

pub fn lock(self) -> LockedBlock

Turn this into a LockedBlock, unable to be reopened again.

pub fn reopen(self, engine: &EthEngine) -> OpenBlock

Given an engine reference, reopen the ClosedBlock into an OpenBlock.