Struct ethcore::block::ExecutedBlock

pub struct ExecutedBlock { /* fields omitted */ }

An internal type for a block's common elements.

Methods

impl ExecutedBlock

fn fields_mut(&mut self) -> BlockRefMut

Get a structure containing individual references to all public fields.

fn fields(&self) -> BlockRef

Get a structure containing individual references to all public fields.