Struct ethcore::block::LockedBlock

pub struct LockedBlock { /* fields omitted */ }

Just like ClosedBlock except that we can't reopen it and it's faster.

We actually store the post-Engine::on_close_block state, unlike in ClosedBlock where it's the pre.

Methods

impl LockedBlock

pub fn strip_receipts_outcomes(&mut self)

Removes outcomes from receipts and updates the receipt root.

This is done after the block is enacted for historical reasons. We allow inconsistency in receipts for some chains if validate_receipts_transition is set to non-zero value, so the check only happens if we detect unmatching root first and then fall back to striped receipts.

pub fn hash(&self) -> H256

Get the hash of the header without seal arguments.

pub fn seal(
self,
engine: &EthEngine,
seal: Vec<Bytes>
) -> Result<SealedBlock, BlockError>

Provide a valid seal in order to turn this into a SealedBlock.

NOTE: This does not check the validity of seal with the engine.

pub fn try_seal(
self,
engine: &EthEngine,
seal: Vec<Bytes>
) -> Result<SealedBlock, (Error, LockedBlock)>

Provide a valid seal in order to turn this into a SealedBlock. This does check the validity of seal with the engine. Returns the ClosedBlock back again if the seal is no good.