Struct ethcore::block_queue::BlockQueue

pub struct BlockQueue {
    // some fields omitted
}

A queue of blocks. Sits between network or other I/O and the BlockChain. Sorts them ready for blockchain insertion.

Methods

impl BlockQueue

fn new(config: BlockQueueConfig, engine: Arc<Box<Engine>>, message_channel: IoChannel<NetSyncMessage>) -> BlockQueue

Creates a new queue instance.

fn clear(&self)

Clear the queue and stop verification activity.

fn flush(&self)

Wait for unverified queue to be empty

fn block_status(&self, hash: &H256) -> BlockStatus

Check if the block is currently in the queue

fn import_block(&self, bytes: Bytes) -> ImportResult

Add a block to the queue.

fn mark_as_bad(&self, block_hashes: &[H256])

Mark given block and all its children as bad. Stops verification.

fn mark_as_good(&self, block_hashes: &[H256])

Mark given block as processed

fn drain(&self, max: usize) -> Vec<PreverifiedBlock>

Removes up to max verified blocks from the queue

fn queue_info(&self) -> BlockQueueInfo

Get queue status.

fn collect_garbage(&self)

Optimise memory footprint of the heap fields.