Struct ethcore::blockchain::BlockChain

pub struct BlockChain {
    // some fields omitted
}

Structure providing fast access to blockchain data.

Does not do input data verification.

Methods

impl BlockChain

fn new(genesis: &[u8], path: &Path) -> BlockChain

Create new instance of blockchain from given Genesis

extern crate ethcore_util as util;
extern crate ethcore;
use std::env;
use std::str::FromStr;
use ethcore::spec::*;
use ethcore::blockchain::*;
use ethcore::ethereum;
use util::hash::*;
use util::uint::*;

fn main() {
    let spec = ethereum::new_frontier();

    let mut dir = env::temp_dir();
    dir.push(H32::random().hex());

    let bc = BlockChain::new(&spec.genesis_block(), &dir);

    let genesis_hash = "d4e56740f876aef8c010b86a40d5f56745a118d0906a34e69aec8c0db1cb8fa3";
    assert_eq!(bc.genesis_hash(), H256::from_str(genesis_hash).unwrap());
    assert!(bc.is_known(&bc.genesis_hash()));
    assert_eq!(bc.genesis_hash(), bc.block_hash(0).unwrap());
}

fn ensure_good(&mut self, _state: &JournalDB)

Ensure that the best block does indeed have a state_root in the state DB. If it doesn't, then rewind down until we find one that does and delete data to ensure that later blocks will be reimported.

fn tree_route(&self, from: H256, to: H256) -> Option<TreeRoute>

Returns a tree route between from and to, which is a tuple of:

a vector of hashes of all blocks, ordered from from to to.

common ancestor of these blocks.

an index where best common ancestor would be

1.) from newer to older

bc: A1 -> A2 -> A3 -> A4 -> A5

from: A5, to: A4

route:

{ blocks: [A5], ancestor: A4, index: 1 }

2.) from older to newer

bc: A1 -> A2 -> A3 -> A4 -> A5

from: A3, to: A4

route:

{ blocks: [A4], ancestor: A3, index: 0 }

3.) fork:

bc:

A1 -> A2 -> A3 -> A4
         -> B3 -> B4

from: B4, to: A4

route:

{ blocks: [B4, B3, A3, A4], ancestor: A2, index: 2 }

fn insert_block(&self, bytes: &[u8])

Inserts the block into backing cache database. Expects the block to be valid and already verified. If the block is already known, does nothing.

fn is_known_transaction(&self, hash: &H256) -> bool

Returns true if transaction is known.

fn best_block_hash(&self) -> H256

Get best block hash.

fn best_block_number(&self) -> BlockNumber

Get best block number.

fn best_block_total_difficulty(&self) -> U256

Get best block total difficulty.

fn log_blooms(&self, hash: &H256) -> Option<BlockLogBlooms>

Get the transactions' log blooms of a block.

fn cache_size(&self) -> CacheSize

Get current cache size.

fn squeeze_to_fit(&self, size: CacheSize)

Tries to squeeze the cache if its too big.

fn collect_garbage(&self, force: bool)

Ticks our cache system and throws out any old data.