Struct ethcore::blockchain::TreeRoute

pub struct TreeRoute {
    pub blocks: Vec<H256>,
    pub ancestor: H256,
    pub index: usize,
}

Represents a tree route between from block and to block:

blocks - a vector of hashes of all blocks, ordered from from to to.

ancestor - best common ancestor of these blocks.

index - an index where best common ancestor would be.

Fields

blocks

TODO [debris] Please document me

ancestor

TODO [debris] Please document me

index

TODO [debris] Please document me