Struct ethcore::builtin::Builtin

pub struct Builtin { /* fields omitted */ }

Pricing scheme, execution definition, and activation block for a built-in contract.

Call cost to compute cost for the given input, execute to execute the contract on the given input, and is_active to determine whether the contract is active.

Unless is_active is true,

Methods

impl Builtin

pub fn cost(&self, input: &[u8]) -> U256

Simple forwarder for cost.

pub fn execute(&self, input: &[u8], output: &mut BytesRef) -> Result<(), Error>

Simple forwarder for execute.

pub fn is_active(&self, at: u64) -> bool

Whether the builtin is activated at the given block number.