Struct ethcore::client::client::Client

pub struct Client {
    // some fields omitted
}

Blockchain database client backed by a persistent database. Owns and manages a blockchain and a block queue. Call import_block() to import a block asynchronously; flush_queue() flushes the queue.

Methods

impl Client

fn new(config: ClientConfig, spec: Spec, path: &Path, miner: Arc<Miner>, message_channel: IoChannel<ClientIoMessage>) -> Result<Arc<Client>, ClientError>

Create a new client with given spec and DB path and custom verifier.

fn set_notify(&self, target: &Arc<ChainNotify>)

Sets the actor to be notified on certain events

fn flush_queue(&self)

Flush the block import queue.

fn import_verified_blocks(&self, io: &IoChannel<ClientIoMessage>) -> usize

This is triggered by a message coming from a block queue when the block is ready for insertion

fn import_queued_transactions(&self, transactions: &[Bytes]) -> usize

Import transactions from the IO queue

fn state_at(&self, id: BlockID) -> Option<State>

Attempt to get a copy of a specific block's state.

This will not fail if given BlockID::Latest. Otherwise, this can fail (but may not) if the DB prunes state.

fn state(&self) -> State

Get a copy of the best block's state.

fn blockchain_cache_info(&self) -> BlockChainCacheSize

Get info on the cache.

fn report(&self) -> ClientReport

Get the report.

fn tick(&self)

Tick the client.

fn configure_cache(&self, pref_cache_size: usize, max_cache_size: usize)

Set up the cache behaviour.

fn block_number(&self, id: BlockID) -> Option<BlockNumber>

Look up the block number for the given block ID.