Struct ethcore::client::EvmTestClient

pub struct EvmTestClient<'a> { /* fields omitted */ }

Simplified, single-block EVM test client.

Methods

impl<'a> EvmTestClient<'a>

pub fn spec_from_json(spec: &ForkSpec) -> Option<&'static Spec>

Converts a json spec definition into spec.

pub fn new(spec: &'a Spec) -> Result<Self, EvmTestError>

Creates new EVM test client with in-memory DB initialized with genesis of given Spec.

pub fn from_pod_state(
spec: &'a Spec,
pod_state: PodState
) -> Result<Self, EvmTestError>

Creates new EVM test client with in-memory DB initialized with given PodState.

pub fn state(&self) -> &State<StateDB>

Return current state.

pub fn call<T: Tracer, V: VMTracer>(
&mut self,
params: ActionParams,
tracer: &mut T,
vm_tracer: &mut V
) -> Result<FinalizationResult, EvmTestError>

Execute the VM given ActionParams and tracer. Returns amount of gas left and the output.

pub fn transact<T: Tracer, V: VMTracer>(
&mut self,
env_info: &EnvInfo,
transaction: SignedTransaction,
tracer: T,
vm_tracer: V
) -> TransactResult<T::Output, V::Output>

Executes a SignedTransaction within context of the provided state and EnvInfo. Returns the state root, gas left and the output.