Struct ethcore::client::Executed

pub struct Executed {
    pub gas: U256,
    pub gas_used: U256,
    pub refunded: U256,
    pub cumulative_gas_used: U256,
    pub logs: Vec<LogEntry>,
    pub contracts_created: Vec<Address>,
    pub output: Bytes,
    pub trace: Option<Trace>,
    pub vm_trace: Option<VMTrace>,
    pub state_diff: Option<StateDiff>,
}

Transaction execution receipt.

Fields

gas

Gas paid up front for execution of transaction.

gas_used

Gas used during execution of transaction.

refunded

Gas refunded after the execution of transaction. To get gas that was required up front, add refunded and gas_used.

cumulative_gas_used

Cumulative gas used in current block so far.

cumulative_gas_used = gas_used(t0) + gas_used(t1) + ... gas_used(tn)

where tn is current transaction.

logs

Vector of logs generated by transaction.

contracts_created

Addresses of contracts created during execution of transaction. Ordered from earliest creation.

eg. sender creates contract A and A in constructor creates contract B

B creation ends first, and it will be the first element of the vector.

output

Transaction output.

trace

The trace of this transaction.

vm_trace

The VM trace of this transaction.

state_diff

The state diff, if we traced it.