Struct ethcore::client::TestBlockChainClient

pub struct TestBlockChainClient {
    pub blocks: RwLock<HashMap<H256, Bytes>>,
    pub numbers: RwLock<HashMap<usize, H256>>,
    pub genesis_hash: H256,
    pub last_hash: RwLock<H256>,
    pub difficulty: RwLock<U256>,
    pub balances: RwLock<HashMap<Address, U256>>,
    pub nonces: RwLock<HashMap<Address, U256>>,
    pub storage: RwLock<HashMap<(Address, H256), H256>>,
    pub code: RwLock<HashMap<Address, Bytes>>,
    pub execution_result: RwLock<Option<Executed>>,
    pub receipts: RwLock<HashMap<TransactionID, LocalizedReceipt>>,
    pub queue_size: AtomicUsize,
    pub miner: Arc<Miner>,
}

Test client.

Fields

blocks

Blocks.

numbers

Mapping of numbers to hashes.

genesis_hash

Genesis block hash.

last_hash

Last block hash.

difficulty

Difficulty.

balances

Balances.

nonces

Nonces.

storage

Storage.

code

Code.

execution_result

Execution result.

receipts

Transaction receipts.

queue_size

Block queue size.

miner

Miner

Methods

impl TestBlockChainClient

fn new() -> Self

Creates new test client.

fn set_transaction_receipt(&self, id: TransactionID, receipt: LocalizedReceipt)

Set the transaction receipt result

fn set_execution_result(&self, result: Executed)

Set the execution result.

fn set_balance(&self, address: Address, balance: U256)

Set the balance of account address to balance.

fn set_nonce(&self, address: Address, nonce: U256)

Set nonce of account address to nonce.

fn set_code(&self, address: Address, code: Bytes)

Set code at address.

fn set_storage(&self, address: Address, position: H256, value: H256)

Set storage position to value for account address.

fn set_queue_size(&self, size: usize)

Set block queue size for testing

fn add_blocks(&self, count: usize, with: EachBlockWith)

Add blocks to test client.

fn corrupt_block(&mut self, n: BlockNumber)

Make a bad block by setting invalid extra data.

fn corrupt_block_parent(&mut self, n: BlockNumber)

Make a bad block by setting invalid parent hash.

fn block_hash_delta_minus(&mut self, delta: usize) -> H256

TODO: