Trait ethcore::client::BlockChainClient

pub trait BlockChainClient: Sync + Send {
    fn block_header(&self, id: BlockID) -> Option<Bytes>;
    fn block_body(&self, id: BlockID) -> Option<Bytes>;
    fn block(&self, id: BlockID) -> Option<Bytes>;
    fn block_status(&self, id: BlockID) -> BlockStatus;
    fn block_total_difficulty(&self, id: BlockID) -> Option<U256>;
    fn nonce(&self, address: &Address, id: BlockID) -> Option<U256>;
    fn block_hash(&self, id: BlockID) -> Option<H256>;
    fn code(&self, address: &Address) -> Option<Bytes>;
    fn balance(&self, address: &Address, id: BlockID) -> Option<U256>;
    fn storage_at(&self, address: &Address, position: &H256, id: BlockID) -> Option<H256>;
    fn transaction(&self, id: TransactionID) -> Option<LocalizedTransaction>;
    fn uncle(&self, id: UncleID) -> Option<Header>;
    fn transaction_receipt(&self, id: TransactionID) -> Option<LocalizedReceipt>;
    fn tree_route(&self, from: &H256, to: &H256) -> Option<TreeRoute>;
    fn find_uncles(&self, hash: &H256) -> Option<Vec<H256>>;
    fn state_data(&self, hash: &H256) -> Option<Bytes>;
    fn block_receipts(&self, hash: &H256) -> Option<Bytes>;
    fn import_block(&self, bytes: Bytes) -> ImportResult;
    fn queue_info(&self) -> BlockQueueInfo;
    fn clear_queue(&self);
    fn chain_info(&self) -> BlockChainInfo;
    fn blocks_with_bloom(&self, bloom: &H2048, from_block: BlockID, to_block: BlockID) -> Option<Vec<BlockNumber>>;
    fn logs(&self, filter: Filter) -> Vec<LocalizedLogEntry>;
    fn call(&self, t: &SignedTransaction, analytics: CallAnalytics) -> Result<Executed, ExecutionError>;
    fn vm_factory(&self) -> &EvmFactory;
    fn filter_traces(&self, filter: TraceFilter) -> Option<Vec<LocalizedTrace>>;
    fn trace(&self, trace: TraceId) -> Option<LocalizedTrace>;
    fn transaction_traces(&self, trace: TransactionID) -> Option<Vec<LocalizedTrace>>;
    fn block_traces(&self, trace: BlockID) -> Option<Vec<LocalizedTrace>>;
    fn last_hashes(&self) -> LastHashes;
    fn import_transactions(&self, transactions: Vec<SignedTransaction>) -> Vec<Result<TransactionImportResult, EthError>>;
    fn queue_transactions(&self, transactions: Vec<Bytes>);
    fn all_transactions(&self) -> Vec<SignedTransaction>;

    fn latest_nonce(&self, address: &Address) -> U256 { ... }
    fn latest_balance(&self, address: &Address) -> U256 { ... }
    fn latest_storage_at(&self, address: &Address, position: &H256) -> H256 { ... }
    fn best_block_header(&self) -> Bytes { ... }
    fn gas_price_statistics(&self, sample_size: usize, distribution_size: usize) -> Result<Vec<U256>, ()> { ... }
    fn dao_rescue_block_gas_limit(&self, chain_hash: H256) -> Option<U256> { ... }
}

Blockchain database client. Owns and manages a blockchain and a block queue.

Required Methods

fn block_header(&self, id: BlockID) -> Option<Bytes>

Get raw block header data by block id.

fn block_body(&self, id: BlockID) -> Option<Bytes>

Get raw block body data by block id. Block body is an RLP list of two items: uncles and transactions.

fn block(&self, id: BlockID) -> Option<Bytes>

Get raw block data by block header hash.

fn block_status(&self, id: BlockID) -> BlockStatus

Get block status by block header hash.

fn block_total_difficulty(&self, id: BlockID) -> Option<U256>

Get block total difficulty.

fn nonce(&self, address: &Address, id: BlockID) -> Option<U256>

Attempt to get address nonce at given block. May not fail on BlockID::Latest.

fn block_hash(&self, id: BlockID) -> Option<H256>

Get block hash.

fn code(&self, address: &Address) -> Option<Bytes>

Get address code.

fn balance(&self, address: &Address, id: BlockID) -> Option<U256>

Get address balance at the given block's state.

May not return None if given BlockID::Latest. Returns None if and only if the block's root hash has been pruned from the DB.

fn storage_at(&self, address: &Address, position: &H256, id: BlockID) -> Option<H256>

Get value of the storage at given position at the given block's state.

May not return None if given BlockID::Latest. Returns None if and only if the block's root hash has been pruned from the DB.

fn transaction(&self, id: TransactionID) -> Option<LocalizedTransaction>

Get transaction with given hash.

fn uncle(&self, id: UncleID) -> Option<Header>

Get uncle with given id.

fn transaction_receipt(&self, id: TransactionID) -> Option<LocalizedReceipt>

Get transaction receipt with given hash.

fn tree_route(&self, from: &H256, to: &H256) -> Option<TreeRoute>

Get a tree route between from and to. See BlockChain::tree_route.

fn find_uncles(&self, hash: &H256) -> Option<Vec<H256>>

Get all possible uncle hashes for a block.

fn state_data(&self, hash: &H256) -> Option<Bytes>

Get latest state node

fn block_receipts(&self, hash: &H256) -> Option<Bytes>

Get raw block receipts data by block header hash.

fn import_block(&self, bytes: Bytes) -> ImportResult

Import a block into the blockchain.

fn queue_info(&self) -> BlockQueueInfo

Get block queue information.

fn clear_queue(&self)

Clear block queue and abort all import activity.

fn chain_info(&self) -> BlockChainInfo

Get blockchain information.

fn blocks_with_bloom(&self, bloom: &H2048, from_block: BlockID, to_block: BlockID) -> Option<Vec<BlockNumber>>

Returns numbers of blocks containing given bloom.

fn logs(&self, filter: Filter) -> Vec<LocalizedLogEntry>

Returns logs matching given filter.

fn call(&self, t: &SignedTransaction, analytics: CallAnalytics) -> Result<Executed, ExecutionError>

Makes a non-persistent transaction call.

fn vm_factory(&self) -> &EvmFactory

Returns EvmFactory.

fn filter_traces(&self, filter: TraceFilter) -> Option<Vec<LocalizedTrace>>

Returns traces matching given filter.

fn trace(&self, trace: TraceId) -> Option<LocalizedTrace>

Returns trace with given id.

fn transaction_traces(&self, trace: TransactionID) -> Option<Vec<LocalizedTrace>>

Returns traces created by transaction.

fn block_traces(&self, trace: BlockID) -> Option<Vec<LocalizedTrace>>

Returns traces created by transaction from block.

fn last_hashes(&self) -> LastHashes

Get last hashes starting from best block.

fn import_transactions(&self, transactions: Vec<SignedTransaction>) -> Vec<Result<TransactionImportResult, EthError>>

import transactions from network/other 3rd party

fn queue_transactions(&self, transactions: Vec<Bytes>)

Queue transactions for importing.

fn all_transactions(&self) -> Vec<SignedTransaction>

list all transactions

Provided Methods

fn latest_nonce(&self, address: &Address) -> U256

Get address nonce at the latest block's state.

fn latest_balance(&self, address: &Address) -> U256

Get address balance at the latest block's state.

fn latest_storage_at(&self, address: &Address, position: &H256) -> H256

Get value of the storage at given position at the latest block's state.

fn best_block_header(&self) -> Bytes

Get the best block header.

fn gas_price_statistics(&self, sample_size: usize, distribution_size: usize) -> Result<Vec<U256>, ()>

Get the gas price distribution.

fn dao_rescue_block_gas_limit(&self, chain_hash: H256) -> Option<U256>

Get Some gas limit of SOFT_FORK_BLOCK, or None if chain is not yet that long.