Trait ethcore::client::traits::Call

pub trait Call {
    type State: StateInfo;
    fn call(
        &self, 
        tx: &SignedTransaction, 
        analytics: CallAnalytics, 
        state: &mut Self::State, 
        header: &Header
    ) -> Result<Executed, CallError>;

    fn call_many(
        &self, 
        txs: &[(SignedTransaction, CallAnalytics)], 
        state: &mut Self::State, 
        header: &Header
    ) -> Result<Vec<Executed>, CallError>;

    fn estimate_gas(
        &self, 
        t: &SignedTransaction, 
        state: &Self::State, 
        header: &Header
    ) -> Result<U256, CallError>;
}

Provides call and call_many methods

Associated Types

type State: StateInfo

Type representing chain state

Required methods

fn call(
&self,
tx: &SignedTransaction,
analytics: CallAnalytics,
state: &mut Self::State,
header: &Header
) -> Result<Executed, CallError>

Makes a non-persistent transaction call.

fn call_many(
&self,
txs: &[(SignedTransaction, CallAnalytics)],
state: &mut Self::State,
header: &Header
) -> Result<Vec<Executed>, CallError>

Makes multiple non-persistent but dependent transaction calls. Returns a vector of successes or a failure if any of the transaction fails.

fn estimate_gas(
&self,
t: &SignedTransaction,
state: &Self::State,
header: &Header
) -> Result<U256, CallError>

Estimates how much gas will be necessary for a call.