Module ethcore::db

Database utilities and definitions.

Enums

CacheUpdatePolicy
Modes for updating caches.

Constants

COL_ACCOUNT_BLOOM
Column for the empty accounts bloom filter.

COL_BODIES
Column for Block bodies

COL_EXTRA
Column for Extras

COL_HEADERS
Column for Block headers

COL_LIGHT_CHAIN
Column for the light client chain.

COL_NODE_INFO
Column for general information from the local node which can persist.

COL_STATE
Column for State

COL_TRACE
Column for Traces

NUM_COLUMNS
Number of columns in DB

Traits

Cache
A cache for arbitrary key-value pairs.

Key
Should be used to get database key associated with given value.

Readable
Should be used to read values from database.

Writable
Should be used to write value into database.