Struct ethcore::encoded::Block

pub struct Block(_);

Owning block view.

Methods

impl Block

pub fn new(raw: Vec<u8>) -> Self

Create a new owning block view. The raw bytes passed in must be an rlp-encoded block.

pub fn new_from_header_and_body(header: &HeaderView, body: &BodyView) -> Self

Create a new owning block view by concatenating the encoded header and body

pub fn view(&self) -> BlockView

Get a borrowed view of the whole block.

pub fn header_view(&self) -> HeaderView

Get a borrowed view of the block header.

pub fn decode(&self) -> Result<FullBlock, DecoderError>

Decode to a full block.

pub fn decode_header(&self) -> FullHeader

Decode the header.

pub fn header(&self) -> Header

Clone the encoded header.

pub fn rlp(&self) -> Rlp

Get the rlp of this block.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn into_inner(self) -> Vec<u8>

Consume the view and return the raw bytes.

impl Block

pub fn hash(&self) -> H256

Returns the header hash.

pub fn parent_hash(&self) -> H256

Returns the parent hash.

pub fn uncles_hash(&self) -> H256

Returns the uncles hash.

pub fn author(&self) -> Address

Returns the author.

pub fn state_root(&self) -> H256

Returns the state root.

pub fn transactions_root(&self) -> H256

Returns the transaction trie root.

pub fn receipts_root(&self) -> H256

Returns the receipts trie root

pub fn log_bloom(&self) -> Bloom

Returns the block log bloom

pub fn difficulty(&self) -> U256

Difficulty of this block

pub fn number(&self) -> BlockNumber

Number of this block.

pub fn timestamp(&self) -> u64

Time this block was produced.

pub fn gas_limit(&self) -> U256

Gas limit of this block.

pub fn gas_used(&self) -> U256

Total gas used in this block.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn extra_data(&self) -> Vec<u8>

Block extra data.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn seal(&self) -> Vec<Vec<u8>>

Engine-specific seal fields.

impl Block

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn transactions(&self) -> Vec<UnverifiedTransaction>

Get a vector of all transactions.

pub fn transactions_count(&self) -> usize

Number of transactions in the block.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn transaction_views(&self) -> Vec<TransactionView>

A view over each transaction in the block.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn transaction_hashes(&self) -> Vec<H256>

The hash of each transaction in the block.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn uncles(&self) -> Vec<FullHeader>

Decode uncle headers.

pub fn uncles_count(&self) -> usize

Number of uncles.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn uncle_views(&self) -> Vec<HeaderView>

Borrowed view over each uncle.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn uncle_hashes(&self) -> Vec<H256>

Hash of each uncle.