Struct ethcore::encoded::Header

pub struct Header(_);

Owning header view.

Methods

impl Header

pub fn new(encoded: Vec<u8>) -> Self

Create a new owning header view. Expects the data to be an RLP-encoded header -- any other case will likely lead to panics further down the line.

pub fn decode(&self) -> Result<FullHeader, DecoderError>

Upgrade this encoded view to a fully owned Header object.

pub fn view(&self) -> HeaderView

Get a borrowed header view onto the data.

pub fn rlp(&self) -> Rlp

Get the rlp of the header.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn into_inner(self) -> Vec<u8>

Consume the view and return the raw bytes.

impl Header

pub fn hash(&self) -> H256

Returns the header hash.

pub fn parent_hash(&self) -> H256

Returns the parent hash.

pub fn uncles_hash(&self) -> H256

Returns the uncles hash.

pub fn author(&self) -> Address

Returns the author.

pub fn state_root(&self) -> H256

Returns the state root.

pub fn transactions_root(&self) -> H256

Returns the transaction trie root.

pub fn receipts_root(&self) -> H256

Returns the receipts trie root

pub fn log_bloom(&self) -> Bloom

Returns the block log bloom

pub fn difficulty(&self) -> U256

Difficulty of this block

pub fn number(&self) -> BlockNumber

Number of this block.

pub fn timestamp(&self) -> u64

Time this block was produced.

pub fn gas_limit(&self) -> U256

Gas limit of this block.

pub fn gas_used(&self) -> U256

Total gas used in this block.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn extra_data(&self) -> Vec<u8>

Block extra data.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn seal(&self) -> Vec<Vec<u8>>

Engine-specific seal fields.