Trait ethcore::engine::Engine

pub trait Engine: Sync + Send {
    fn name(&self) -> &str;
    fn spec(&self) -> &Spec;
    fn vm_factory(&self) -> &Factory;
    fn schedule(&self, env_info: &EnvInfo) -> Schedule;

    fn version(&self) -> SemanticVersion { ... }
    fn seal_fields(&self) -> usize { ... }
    fn seal_rlp(&self) -> Bytes { ... }
    fn extra_info(&self, _header: &Header) -> HashMap<String, String> { ... }
    fn maximum_extra_data_size(&self) -> usize { ... }
    fn maximum_uncle_count(&self) -> usize { ... }
    fn account_start_nonce(&self) -> U256 { ... }
    fn on_new_block(&self, _block: &mut Block) { ... }
    fn on_close_block(&self, _block: &mut Block) { ... }
    fn verify_block_basic(&self, _header: &Header, _block: Option<&[u8]>) -> Result<(), Error> { ... }
    fn verify_block_unordered(&self, _header: &Header, _block: Option<&[u8]>) -> Result<(), Error> { ... }
    fn verify_block_family(&self, _header: &Header, _parent: &Header, _block: Option<&[u8]>) -> Result<(), Error> { ... }
    fn verify_transaction_basic(&self, _t: &Transaction, _header: &Header) -> Result<(), Error> { ... }
    fn verify_transaction(&self, _t: &Transaction, _header: &Header) -> Result<(), Error> { ... }
    fn populate_from_parent(&self, _header: &mut Header, _parent: &Header) { ... }
    fn is_builtin(&self, a: &Address) -> bool { ... }
    fn cost_of_builtin(&self, a: &Address, input: &[u8]) -> U256 { ... }
    fn execute_builtin(&self, a: &Address, input: &[u8], output: &mut [u8]) { ... }
}

A consensus mechanism for the chain. Generally either proof-of-work or proof-of-stake-based. Provides hooks into each of the major parts of block import.

Required Methods

fn name(&self) -> &str

The name of this engine.

fn spec(&self) -> &Spec

Get the general parameters of the chain.

fn vm_factory(&self) -> &Factory

Get current EVM factory

fn schedule(&self, env_info: &EnvInfo) -> Schedule

Get the EVM schedule for the given env_info.

Provided Methods

fn version(&self) -> SemanticVersion

The version of this engine. Should be of the form

fn seal_fields(&self) -> usize

The number of additional header fields required for this engine.

fn seal_rlp(&self) -> Bytes

Default values of the additional fields RLP-encoded in a raw (non-list) harness.

fn extra_info(&self, _header: &Header) -> HashMap<String, String>

Additional engine-specific information for the user/developer concerning header.

fn maximum_extra_data_size(&self) -> usize

Some intrinsic operation parameters; by default they take their value from the spec()'s engine_params.

fn maximum_uncle_count(&self) -> usize

TODO [Gav Wood] Please document me

fn account_start_nonce(&self) -> U256

TODO [Gav Wood] Please document me

fn on_new_block(&self, _block: &mut Block)

Block transformation functions, before and after the transactions.

fn on_close_block(&self, _block: &mut Block)

TODO [Gav Wood] Please document me

fn verify_block_basic(&self, _header: &Header, _block: Option<&[u8]>) -> Result<(), Error>

Phase 1 quick block verification. Only does checks that are cheap. block (the header's full block) may be provided for additional checks. Returns either a null Ok or a general error detailing the problem with import.

fn verify_block_unordered(&self, _header: &Header, _block: Option<&[u8]>) -> Result<(), Error>

Phase 2 verification. Perform costly checks such as transaction signatures. block (the header's full block) may be provided for additional checks. Returns either a null Ok or a general error detailing the problem with import.

fn verify_block_family(&self, _header: &Header, _parent: &Header, _block: Option<&[u8]>) -> Result<(), Error>

Phase 3 verification. Check block information against parent and uncles. block (the header's full block) may be provided for additional checks. Returns either a null Ok or a general error detailing the problem with import.

fn verify_transaction_basic(&self, _t: &Transaction, _header: &Header) -> Result<(), Error>

Additional verification for transactions in blocks.

fn verify_transaction(&self, _t: &Transaction, _header: &Header) -> Result<(), Error>

TODO [Gav Wood] Please document me

fn populate_from_parent(&self, _header: &mut Header, _parent: &Header)

Don't forget to call Super::populateFromParent when subclassing & overriding.

fn is_builtin(&self, a: &Address) -> bool

TODO [Gav Wood] Please document me

fn cost_of_builtin(&self, a: &Address, input: &[u8]) -> U256

TODO [Gav Wood] Please document me

fn execute_builtin(&self, a: &Address, input: &[u8], output: &mut [u8])

TODO [Gav Wood] Please document me