Module ethcore::engines::block_reward

A module with types for declaring block rewards and a client interface for interacting with a block reward contract.

Modules

block_reward_contract

Structs

BlockRewardContract
A client for the block reward contract.

Enums

RewardKind
The kind of block reward. Depending on the consensus engine the allocated block reward might have different semantics which could lead e.g. to different reward values.

Functions

apply_block_rewards
Applies the given block rewards, i.e. adds the given balance to each benefactors' address. If tracing is enabled the operations are recorded.