Enum ethcore::engines::EngineError

pub enum EngineError {
    NotAuthorized(Address),
    DoubleVote(Address),
    NotProposer(Mismatch<Address>),
    UnexpectedMessage,
    BadSealFieldSize(OutOfBounds<usize>),
    InsufficientProof(String),
    FailedSystemCall(String),
    MalformedMessage(String),
    RequiresClient,
}

Voting errors.

Variants
NotAuthorized(Address)

Signature or author field does not belong to an authority.

DoubleVote(Address)

The same author issued different votes at the same step.

NotProposer(Mismatch<Address>)

The received block is from an incorrect proposer.

UnexpectedMessage

Message was not expected.

BadSealFieldSize(OutOfBounds<usize>)

Seal field has an unexpected size.

InsufficientProof(String)

Validation proof insufficient.

FailedSystemCall(String)

Failed system call.

MalformedMessage(String)

Malformed consensus message.

RequiresClient

Requires client ref, but none registered.