Module ethcore::error

General error types for use in ethcore.

Re-exports

pub use executed::CallError;

pub use executed::ExecutionError;

Structs

Error

The Error type.

ImportError

The Error type.

QueueError

The Error type.

Enums

BlockError

Errors concerning block processing.

ErrorKind

The kind of an error.

ImportErrorKind

The kind of an error.

QueueErrorKind

The kind of an error.

TransactionImportError

Api-level error for transaction import

Traits

ErrorResultExt

Additional methods for Result, for easy interaction with this crate.

ImportErrorResultExt

Additional methods for Result, for easy interaction with this crate.

QueueErrorResultExt

Additional methods for Result, for easy interaction with this crate.

Type Definitions

EthcoreResult

Convenient wrapper around std::Result.

ImportErrorResult

Convenient wrapper around std::Result.

QueueErrorResult

Convenient wrapper around std::Result.