Enum ethcore::error::BlockError

pub enum BlockError {
    TooManyUncles(OutOfBounds<usize>),
    ExtraDataOutOfBounds(OutOfBounds<usize>),
    InvalidSealArity(Mismatch<usize>),
    TooMuchGasUsed(OutOfBounds<U256>),
    InvalidUnclesHash(Mismatch<H256>),
    UncleTooOld(OutOfBounds<BlockNumber>),
    UncleIsBrother(OutOfBounds<BlockNumber>),
    UncleInChain(H256),
    UncleParentNotInChain(H256),
    InvalidStateRoot(Mismatch<H256>),
    InvalidGasUsed(Mismatch<U256>),
    InvalidTransactionsRoot(Mismatch<H256>),
    DifficultyOutOfBounds(OutOfBounds<U256>),
    InvalidDifficulty(Mismatch<U256>),
    MismatchedH256SealElement(Mismatch<H256>),
    InvalidProofOfWork(OutOfBounds<U256>),
    InvalidSeal,
    InvalidGasLimit(OutOfBounds<U256>),
    InvalidReceiptsRoot(Mismatch<H256>),
    InvalidTimestamp(OutOfBounds<u64>),
    InvalidLogBloom(Mismatch<LogBloom>),
    InvalidParentHash(Mismatch<H256>),
    InvalidNumber(Mismatch<BlockNumber>),
    RidiculousNumber(OutOfBounds<BlockNumber>),
    UnknownParent(H256),
    UnknownUncleParent(H256),
}

Errors concerning block processing.

Variants

TooManyUncles

Block has too many uncles.

ExtraDataOutOfBounds

Extra data is of an invalid length.

InvalidSealArity

Seal is incorrect format.

TooMuchGasUsed

Block has too much gas used.

InvalidUnclesHash

Uncles hash in header is invalid.

UncleTooOld

An uncle is from a generation too old.

UncleIsBrother

An uncle is from the same generation as the block.

UncleInChain

An uncle is already in the chain.

UncleParentNotInChain

An uncle has a parent not in the chain.

InvalidStateRoot

State root header field is invalid.

InvalidGasUsed

Gas used header field is invalid.

InvalidTransactionsRoot

Transactions root header field is invalid.

DifficultyOutOfBounds

Difficulty is out of range; this can be used as an looser error prior to getting a definitive value for difficulty. This error needs only provide bounds of which it is out.

InvalidDifficulty

Difficulty header field is invalid; this is a strong error used after getting a definitive value for difficulty (which is provided).

MismatchedH256SealElement

Seal element of type H256 (max_hash for Ethash, but could be something else for other seal engines) is out of bounds.

InvalidProofOfWork

Proof-of-work aspect of seal, which we assume is a 256-bit value, is invalid.

InvalidSeal

Some low-level aspect of the seal is incorrect.

InvalidGasLimit

Gas limit header field is invalid.

InvalidReceiptsRoot

Receipts trie root header field is invalid.

InvalidTimestamp

Timestamp header field is invalid.

InvalidLogBloom

Log bloom header field is invalid.

InvalidParentHash

Parent hash field of header is invalid; this is an invalid error indicating a logic flaw in the codebase. TODO: remove and favour an assert!/panic!.

InvalidNumber

Number field of header is invalid.

RidiculousNumber

Block number isn't sensible.

UnknownParent

Parent given is unknown.

UnknownUncleParent

Uncle parent given is unknown.