Enum ethcore::error::Error

pub enum Error {
    Client(ClientError),
    Util(UtilError),
    Block(BlockError),
    UnknownEngineName(String),
    Execution(ExecutionError),
    Transaction(TransactionError),
    Import(ImportError),
    PowHashInvalid,
    PowInvalid,
    Trie(TrieError),
}

General error type which should be capable of representing all errors in ethcore.

Variants

Client

Client configuration error.

Util

Error concerning a utility.

Block

Error concerning block processing.

UnknownEngineName

Unknown engine given.

Execution

Error concerning EVM code execution.

Transaction

Error concerning transaction processing.

Import

Error concerning block import.

PowHashInvalid

PoW hash is invalid or out of date.

PowInvalid

The value of the nonce or mishash is invalid.

Trie

Error concerning TrieDBs