Enum ethcore::error::ExecutionError

pub enum ExecutionError {
    NotEnoughBaseGas {
        required: U256,
        got: U256,
    },
    BlockGasLimitReached {
        gas_limit: U256,
        gas_used: U256,
        gas: U256,
    },
    InvalidNonce {
        expected: U256,
        got: U256,
    },
    NotEnoughCash {
        required: U512,
        got: U512,
    },
    Internal,
}

Result of executing the transaction.

Variants

NotEnoughBaseGas

Returned when there gas paid for transaction execution is lower than base gas required. TODO [Gav Wood] Please document me

Fields

required

TODO [Gav Wood] Please document me

got

TODO [Gav Wood] Please document me

BlockGasLimitReached

Returned when block (gas_used + gas) > gas_limit.

If gas =< gas_limit, upstream may try to execute the transaction in next block.

Fields

gas_limit

TODO [Gav Wood] Please document me

gas_used

TODO [Gav Wood] Please document me

gas

TODO [Gav Wood] Please document me

InvalidNonce

Returned when transaction nonce does not match state nonce.

Fields

expected

TODO [Gav Wood] Please document me

got

TODO [Gav Wood] Please document me

NotEnoughCash

Returned when cost of transaction (value + gas_price * gas) exceeds current sender balance.

Fields

required

TODO [Gav Wood] Please document me

got

TODO [Gav Wood] Please document me

Internal

Returned when internal evm error occurs.