Enum ethcore::error::TransactionError

pub enum TransactionError {
    AlreadyImported,
    Old,
    TooCheapToReplace,
    LimitReached,
    InsufficientGasPrice {
        minimal: U256,
        got: U256,
    },
    InsufficientBalance {
        balance: U256,
        cost: U256,
    },
    GasLimitExceeded {
        limit: U256,
        got: U256,
    },
    InvalidGasLimit(OutOfBounds<U256>),
    DAORescue,
}

Errors concerning transaction processing.

Variants

AlreadyImported

Transaction is already imported to the queue

Old

Transaction is not valid anymore (state already has higher nonce)

TooCheapToReplace

Transaction has too low fee (there is already a transaction with the same sender-nonce but higher gas price)

LimitReached

Transaction was not imported to the queue because limit has been reached.

InsufficientGasPrice

Transaction's gas price is below threshold.

Fields

minimal

Minimal expected gas price

got

Transaction gas price

InsufficientBalance

Sender doesn't have enough funds to pay for this transaction

Fields

balance

Senders balance

cost

Transaction cost

GasLimitExceeded

Transactions gas is higher then current gas limit

Fields

limit

Current gas limit

got

Declared transaction gas

InvalidGasLimit

Transaction's gas limit (aka gas) is invalid.

DAORescue

Transaction is invalid for some other reason.