Module ethcore::ethereum

Ethereum protocol module.

Contains all Ethereum network specific stuff, such as denominations and consensus specifications.

Re-exports

pub use self::ethash::Ethash;

pub use self::denominations::*;

Modules

denominations
Export the denominations module.

ethash
Export the ethash module.

Functions

new_byzantium_test
Create a new Foundation Byzantium era spec.

new_byzantium_test_machine
Create a new Foundation Byzantium era spec.

new_classic
Create a new Classic Mainnet chain spec without the DAO hardfork.

new_constantinople_test
Create a new Foundation Constantinople era spec.

new_constantinople_test_machine
Create a new Foundation Constantinople era spec.

new_easthub
Create a new Easthub mainnet chain spec.

new_eip150_test
Create a new Foundation Homestead-EIP150-era chain spec as though it never changed from Homestead/Frontier.

new_eip161_test
Create a new Foundation Homestead-EIP161-era chain spec as though it never changed from Homestead/Frontier.

new_ellaism
Create a new Ellaism mainnet chain spec.

new_expanse
Create a new Expanse mainnet chain spec.

new_foundation
Create a new Foundation Mainnet chain spec.

new_frontier_test
Create a new Foundation Frontier-era chain spec as though it never changes to Homestead.

new_frontier_test_machine
Create a new Foundation Frontier-era chain spec as though it never changes to Homestead.

new_homestead_test
Create a new Foundation Homestead-era chain spec as though it never changed from Frontier.

new_homestead_test_machine
Create a new Foundation Homestead-era chain spec as though it never changed from Frontier.

new_kovan
Create a new Kovan testnet chain spec.

new_kovan_wasm_test_machine
Create new Kovan spec with wasm activated at certain block

new_mainnet_like
Create a new Foundation Mainnet chain spec without genesis accounts.

new_mcip3_test
Create a new Musicoin-MCIP3-era spec.

new_mcip3_test_machine
Create a new Musicoin-MCIP3-era spec.

new_morden
Create a new Morden chain spec.

new_musicoin
Create a new Musicoin mainnet chain spec.

new_olympic
Create a new Foundation Olympic chain spec.

new_ropsten
Create a new Foundation Ropsten chain spec.

new_ropsten_test
Create a new Ropsten chain spec.

new_social
Create a new Ethereum Social mainnet chain spec.

new_tobalaba
Create a new Tobalaba chain spec.

new_transition_test
Create a new Foundation Frontier/Homestead/DAO chain spec with transition points at #5 and #8.