Struct ethcore::ethereum::ethash::EthashParams

pub struct EthashParams {
    pub minimum_difficulty: U256,
    pub difficulty_bound_divisor: U256,
    pub difficulty_increment_divisor: u64,
    pub metropolis_difficulty_increment_divisor: u64,
    pub duration_limit: u64,
    pub homestead_transition: u64,
    pub difficulty_hardfork_transition: u64,
    pub difficulty_hardfork_bound_divisor: U256,
    pub bomb_defuse_transition: u64,
    pub eip100b_transition: u64,
    pub ecip1010_pause_transition: u64,
    pub ecip1010_continue_transition: u64,
    pub ecip1017_era_rounds: u64,
    pub mcip3_transition: u64,
    pub mcip3_miner_reward: U256,
    pub mcip3_ubi_reward: U256,
    pub mcip3_ubi_contract: Address,
    pub mcip3_dev_reward: U256,
    pub mcip3_dev_contract: Address,
    pub block_reward: U256,
    pub eip649_transition: u64,
    pub eip649_delay: u64,
    pub eip649_reward: Option<U256>,
    pub expip2_transition: u64,
    pub expip2_duration_limit: u64,
}

Ethash params.

Fields
minimum_difficulty: U256

Minimum difficulty.

difficulty_bound_divisor: U256

Difficulty bound divisor.

difficulty_increment_divisor: u64

Difficulty increment divisor.

metropolis_difficulty_increment_divisor: u64

Metropolis difficulty increment divisor.

duration_limit: u64

Block duration.

homestead_transition: u64

Homestead transition block number.

difficulty_hardfork_transition: u64

Transition block for a change of difficulty params (currently just bound_divisor).

difficulty_hardfork_bound_divisor: U256

Difficulty param after the difficulty transition.

bomb_defuse_transition: u64

Block on which there is no additional difficulty from the exponential bomb.

eip100b_transition: u64

Number of first block where EIP-100 rules begin.

ecip1010_pause_transition: u64

Number of first block where ECIP-1010 begins.

ecip1010_continue_transition: u64

Number of first block where ECIP-1010 ends.

ecip1017_era_rounds: u64

Total block number for one ECIP-1017 era.

mcip3_transition: u64

Number of first block where MCIP-3 begins.

mcip3_miner_reward: U256

MCIP-3 Block reward coin-base for miners.

mcip3_ubi_reward: U256

MCIP-3 Block reward ubi-base for basic income.

mcip3_ubi_contract: Address

MCIP-3 contract address for universal basic income.

mcip3_dev_reward: U256

MCIP-3 Block reward dev-base for dev funds.

mcip3_dev_contract: Address

MCIP-3 contract address for the developer funds.

block_reward: U256

Block reward in base units.

eip649_transition: u64

EIP-649 transition block.

eip649_delay: u64

EIP-649 bomb delay.

eip649_reward: Option<U256>

EIP-649 base reward.

expip2_transition: u64

EXPIP-2 block height

expip2_duration_limit: u64

EXPIP-2 duration limit