Module ethcore::evm::evm

Evm interface.

Enums

Error
Evm errors.

GasLeft
Gas Left: either it is a known value, or it needs to be computed by processing a return instruction.

Traits

CostType
Cost calculation type. For low-gas usage we calculate costs using usize instead of U256

Evm
Evm interface

Finalize
Types that can be "finalized" using an EVM.

Type Definitions

Result
A specialized version of Result over EVM errors.