Enum ethcore::evm::evm::Error

pub enum Error {
    OutOfGas,
    BadJumpDestination {
        destination: usize,
    },
    BadInstruction {
        instruction: u8,
    },
    StackUnderflow {
        instruction: &'static str,
        wanted: usize,
        on_stack: usize,
    },
    OutOfStack {
        instruction: &'static str,
        wanted: usize,
        limit: usize,
    },
    Internal,
}

Evm errors.

Variants

OutOfGas

OutOfGas is returned when transaction execution runs out of gas. The state should be reverted to the state from before the transaction execution. But it does not mean that transaction was invalid. Balance still should be transfered and nonce should be increased.

BadJumpDestination

BadJumpDestination is returned when execution tried to move to position that wasn't marked with JUMPDEST instruction

Fields

destination

TODO [Tomusdrw] Please document me

BadInstruction

BadInstructions is returned when given instruction is not supported

Fields

instruction

TODO [Tomusdrw] Please document me

StackUnderflow

StackUnderflow when there is not enough stack elements to execute instruction First parameter says how many elements were needed and the second how many were actually on Stack

Fields

instruction

TODO [Tomusdrw] Please document me

wanted

TODO [Tomusdrw] Please document me

on_stack

TODO [Tomusdrw] Please document me

OutOfStack

When execution would exceed defined Stack Limit

Fields

instruction

TODO [Tomusdrw] Please document me

wanted

TODO [Tomusdrw] Please document me

limit

TODO [Tomusdrw] Please document me

Internal

Returned on evm internal error. Should never be ignored during development. Likely to cause consensus issues.