Struct ethcore::evm::schedule::Schedule

pub struct Schedule {
    pub exceptional_failed_code_deposit: bool,
    pub have_delegate_call: bool,
    pub stack_limit: usize,
    pub max_depth: usize,
    pub tier_step_gas: [usize; 8],
    pub exp_gas: usize,
    pub exp_byte_gas: usize,
    pub sha3_gas: usize,
    pub sha3_word_gas: usize,
    pub sload_gas: usize,
    pub sstore_set_gas: usize,
    pub sstore_reset_gas: usize,
    pub sstore_refund_gas: usize,
    pub jumpdest_gas: usize,
    pub log_gas: usize,
    pub log_data_gas: usize,
    pub log_topic_gas: usize,
    pub create_gas: usize,
    pub call_gas: usize,
    pub call_stipend: usize,
    pub call_value_transfer_gas: usize,
    pub call_new_account_gas: usize,
    pub suicide_refund_gas: usize,
    pub memory_gas: usize,
    pub quad_coeff_div: usize,
    pub create_data_gas: usize,
    pub tx_gas: usize,
    pub tx_create_gas: usize,
    pub tx_data_zero_gas: usize,
    pub tx_data_non_zero_gas: usize,
    pub copy_gas: usize,
}

Definition of the cost schedule and other parameterisations for the EVM.

Fields

exceptional_failed_code_deposit

Does it support exceptional failed code deposit

have_delegate_call

Does it have a delegate cal

stack_limit

VM stack limit

max_depth

Max number of nested calls/creates

tier_step_gas

Gas prices for instructions in all tiers

exp_gas

Gas price for EXP opcode

exp_byte_gas

Additional gas for EXP opcode for each byte of exponent

sha3_gas

Gas price for SHA3 opcode

sha3_word_gas

Additional gas for SHA3 opcode for each word of hashed memory

sload_gas

Gas price for loading from storage

sstore_set_gas

Gas price for setting new value to storage (storage==0, new!=0)

sstore_reset_gas

Gas price for altering value in storage

sstore_refund_gas

Gas refund for SSTORE clearing (when storage!=0, new==0)

jumpdest_gas

Gas price for JUMPDEST opcode

log_gas

Gas price for LOG*

log_data_gas

Additional gas for data in LOG*

log_topic_gas

Additional gas for each topic in LOG*

create_gas

Gas price for CREATE opcode

call_gas

Gas price for *CALL* opcodes

call_stipend

Stipend for transfer for CALL|CALLCODE opcode when value>0

call_value_transfer_gas

Additional gas required for value transfer (CALL|CALLCODE)

call_new_account_gas

Additional gas for creating new account (CALL|CALLCODE)

suicide_refund_gas

Refund for SUICIDE

memory_gas

Gas for used memory

quad_coeff_div

Coefficient used to convert memory size to gas price for memory

create_data_gas

Cost for contract length when executing CREATE

tx_gas

Transaction cost

tx_create_gas

CREATE transaction cost

tx_data_zero_gas

Additional cost for empty data transaction

tx_data_non_zero_gas

Aditional cost for non-empty data transaction

copy_gas

Gas price for copying memory

Methods

impl Schedule

fn new_frontier() -> Schedule

Schedule for the Frontier-era of the Ethereum main net.

fn new_homestead() -> Schedule

Schedule for the Homestead-era of the Ethereum main net.