Struct ethcore::header::Header

pub struct Header {
    pub parent_hash: H256,
    pub timestamp: u64,
    pub number: BlockNumber,
    pub author: Address,
    pub transactions_root: H256,
    pub uncles_hash: H256,
    pub extra_data: Bytes,
    pub state_root: H256,
    pub receipts_root: H256,
    pub log_bloom: LogBloom,
    pub gas_used: U256,
    pub gas_limit: U256,
    pub difficulty: U256,
    pub seal: Vec<Bytes>,
    pub hash: RefCell<Option<H256>>,
    pub bare_hash: RefCell<Option<H256>>,
}

A block header.

Reflects the specific RLP fields of a block in the chain with additional room for the seal which is non-specific.

Doesn't do all that much on its own.

Fields

parent_hash

Parent hash.

timestamp

Block timestamp.

number

Block number.

author

Block author.

transactions_root

Transactions root.

uncles_hash

Block uncles hash.

extra_data

Block extra data.

state_root

State root.

receipts_root

Block receipts root.

log_bloom

Block bloom.

gas_used

Gas used for contracts execution.

gas_limit

Block gas limit.

difficulty

Block difficulty.

seal

Vector of post-RLP-encoded fields.

hash

The memoized hash of the RLP representation including the seal fields.

bare_hash

The memoized hash of the RLP representation without the seal fields.

Methods

impl Header

fn signature(&self) -> H520

Get the none field of the header.

impl Header

fn nonce(&self) -> H64

Get the none field of the header.

fn mix_hash(&self) -> H256

Get the mix hash field of the header.

fn set_nonce_and_mix_hash(&mut self, nonce: &H64, mix_hash: &H256)

Set the nonce and mix hash fields of the header.

impl Header

fn new() -> Self

Create a new, default-valued, header.

fn parent_hash(&self) -> &H256

Get the parent_hash field of the header.

fn timestamp(&self) -> u64

Get the timestamp field of the header.

fn number(&self) -> BlockNumber

Get the number field of the header.

fn author(&self) -> &Address

Get the author field of the header.

fn extra_data(&self) -> &Bytes

Get the extra data field of the header.

fn state_root(&self) -> &H256

Get the state root field of the header.

fn receipts_root(&self) -> &H256

Get the receipts root field of the header.

fn gas_limit(&self) -> &U256

Get the gas limit field of the header.

fn difficulty(&self) -> &U256

Get the difficulty field of the header.

fn seal(&self) -> &[Bytes]

Get the seal field of the header.

fn set_parent_hash(&mut self, a: H256)

Set the number field of the header.

fn set_uncles_hash(&mut self, a: H256)

Set the uncles hash field of the header.

fn set_state_root(&mut self, a: H256)

Set the state root field of the header.

fn set_transactions_root(&mut self, a: H256)

Set the transactions root field of the header.

fn set_receipts_root(&mut self, a: H256)

Set the receipts root field of the header.

fn set_log_bloom(&mut self, a: LogBloom)

Set the log bloom field of the header.

fn set_timestamp(&mut self, a: u64)

Set the timestamp field of the header.

fn set_timestamp_now(&mut self, but_later_than: u64)

Set the timestamp field of the header to the current time.

fn set_number(&mut self, a: BlockNumber)

Set the number field of the header.

fn set_author(&mut self, a: Address)

Set the author field of the header.

fn set_extra_data(&mut self, a: Bytes)

Set the extra data field of the header.

fn set_gas_used(&mut self, a: U256)

Set the gas used field of the header.

fn set_gas_limit(&mut self, a: U256)

Set the gas limit field of the header.

fn set_difficulty(&mut self, a: U256)

Set the difficulty field of the header.

fn set_seal(&mut self, a: Vec<Bytes>)

Set the seal field of the header.

fn hash(&self) -> H256

Get the hash of this header (sha3 of the RLP).

fn bare_hash(&self) -> H256

Get the hash of the header excluding the seal

fn note_dirty(&self)

Note that some fields have changed. Resets the memoised hash.

fn stream_rlp(&self, s: &mut RlpStream, with_seal: Seal)

Place this header into an RLP stream s, optionally with_seal.

fn rlp(&self, with_seal: Seal) -> Bytes

Get the RLP of this header, optionally with_seal.

fn rlp_sha3(&self, with_seal: Seal) -> H256

Get the SHA3 (Keccak) of this header, optionally with_seal.