Module ethcore::machine

Ethereum-like state machine definition.

Structs

AuxiliaryData
Auxiliary data fetcher for an Ethereum machine. In Ethereum-like machines there are two kinds of auxiliary data: bodies and receipts.

EthashExtensions
Ethash-specific extensions.

EthereumMachine
An ethereum-like state machine.

Enums

AuxiliaryRequest
Request for auxiliary data of a block.

Constants

PARITY_GAS_LIMIT_DETERMINANT
Parity tries to round block.gas_limit to multiple of this constant

Traits

WithRewards
A state machine that uses block rewards.

Type Definitions

Call
Type alias for a function we can make calls through synchronously. Returns the call result and state proof for each call.

ScheduleCreationRules
Special rules to be applied to the schedule.