Struct ethcore::machine::EthereumMachine

pub struct EthereumMachine { /* fields omitted */ }

An ethereum-like state machine.

Methods

impl EthereumMachine

pub fn regular(
params: CommonParams,
builtins: BTreeMap<Address, Builtin>
) -> EthereumMachine

Regular ethereum machine.

pub fn with_ethash_extensions(
params: CommonParams,
builtins: BTreeMap<Address, Builtin>,
extensions: EthashExtensions
) -> EthereumMachine

Ethereum machine with ethash extensions.

pub fn set_schedule_creation_rules(&mut self, rules: Box<ScheduleCreationRules>)

Attach special rules to the creation of schedule.

pub fn ethash_extensions(&self) -> Option<&EthashExtensions>

Get a reference to the ethash-specific extensions.

impl EthereumMachine

pub fn execute_as_system(
&self,
block: &mut ExecutedBlock,
contract_address: Address,
gas: U256,
data: Option<Vec<u8>>
) -> Result<Vec<u8>, Error>

Execute a call as the system address. Block environment information passed to the VM is modified to have its gas limit bounded at the upper limit of possible used gases including this system call, capped at the maximum value able to be represented by U256. This system call modifies the block state, but discards other information. If suicides, logs or refunds happen within the system call, they will not be executed or recorded. Gas used by this system call will not be counted on the block.

pub fn on_new_block(&self, block: &mut ExecutedBlock) -> Result<(), Error>

Logic to perform on a new block: updating last hashes and the DAO fork, for ethash.

pub fn populate_from_parent(
&self,
header: &mut Header,
parent: &Header,
gas_floor_target: U256,
gas_ceil_target: U256
)

Populate a header's fields based on its parent's header. Usually implements the chain scoring rule based on weight. The gas floor target must not be lower than the engine's minimum gas limit.

pub fn params(&self) -> &CommonParams

Get the general parameters of the chain.

pub fn schedule(&self, block_number: BlockNumber) -> Schedule

Get the EVM schedule for the given block number.

pub fn builtins(&self) -> &BTreeMap<Address, Builtin>

Builtin-contracts for the chain..

pub fn builtin(
&self,
a: &Address,
block_number: BlockNumber
) -> Option<&Builtin>

Attempt to get a handle to a built-in contract. Only returns references to activated built-ins.

pub fn maximum_extra_data_size(&self) -> usize

Some intrinsic operation parameters; by default they take their value from the spec()'s engine_params.

pub fn account_start_nonce(&self, block: u64) -> U256

The nonce with which accounts begin at given block.

pub fn signing_chain_id(&self, env_info: &EnvInfo) -> Option<u64>

The network ID that transactions should be signed with.

pub fn create_address_scheme(
&self,
number: BlockNumber
) -> CreateContractAddress

Returns new contract address generation scheme at given block number.

pub fn verify_transaction_unordered(
&self,
t: UnverifiedTransaction,
_header: &Header
) -> Result<SignedTransaction, Error>

Verify a particular transaction is valid, regardless of order.

pub fn verify_transaction_basic(
&self,
t: &UnverifiedTransaction,
header: &Header
) -> Result<(), Error>

Does basic verification of the transaction.

pub fn verify_transaction<C: BlockInfo + CallContract>(
&self,
t: &SignedTransaction,
header: &Header,
client: &C
) -> Result<(), Error>

Does verification of the transaction against the parent state.

pub fn additional_params(&self) -> HashMap<String, String>

Additional params.

pub fn decode_transaction(
&self,
transaction: &[u8]
) -> Result<UnverifiedTransaction, Error>

Performs pre-validation of RLP decoded transaction before other processing