Module ethcore::miner

Miner module Keeps track of transactions and currently sealed pending block.

Modules

pool_client
Blockchain access for transaction pool.

stratum
Client-side stratum job dispatcher and mining notifier handler

Structs

AuthoringParams
Configurable parameters of block authoring.

Miner
Keeps track of transactions using priority queue and holds currently mined block. Handles preparing work for "work sealing" or seals "internally" if Engine does not require work.

MinerOptions
Configures the behaviour of the miner.

Enums

Penalization
Transaction queue penalization settings.

PendingOrdering
Transaction ordering when requesting pending set.

PendingSet
Different possible definitions for pending transaction set.

Traits

BlockChainClient
Extended client interface used for mining

MinerService
Miner client API

TransactionVerifierClient
Provides methods to verify incoming external transactions