Enum ethcore::miner::GasPricer

pub enum GasPricer {
    Fixed(U256),
    Calibrated(GasPriceCalibrator),
}

Struct to look after updating the acceptable gas price of a miner.

Variants

Fixed

A fixed gas price in terms of Wei - always the argument given.

Calibrated

Gas price is calibrated according to a fixed amount of USD.

Methods

impl GasPricer

fn new_calibrated(options: GasPriceCalibratorOptions) -> GasPricer

Create a new Calibrated GasPricer.

fn new_fixed(gas_price: U256) -> GasPricer

Create a new Fixed GasPricer.