Enum ethcore::miner::LocalTransactionStatus

pub enum LocalTransactionStatus {
    Pending,
    Future,
    Mined(SignedTransaction),
    Dropped(SignedTransaction),
    Replaced(SignedTransaction, U256, H256),
    Rejected(SignedTransaction, TransactionError),
    Invalid(SignedTransaction),
}

Status of local transaction. Can indicate that the transaction is currently part of the queue (Pending/Future) or gives a reason why the transaction was removed.

Variants
Pending

The transaction is currently in the transaction queue.

Future

The transaction is in future part of the queue.

Mined(SignedTransaction)

Transaction is already mined.

Dropped(SignedTransaction)

Transaction is dropped because of limit

Replaced(SignedTransaction, U256, H256)

Replaced because of higher gas price of another transaction.

Rejected(SignedTransaction, TransactionError)

Transaction was never accepted to the queue.

Invalid(SignedTransaction)

Transaction is invalid.