Enum ethcore::miner::PrioritizationStrategy

pub enum PrioritizationStrategy {
    GasPriceOnly,
    GasAndGasPrice,
    GasFactorAndGasPrice,
}

Describes the strategy used to prioritize transactions in the queue.

Variants
GasPriceOnly

Use only gas price. Disregards the actual computation cost of the transaction. i.e. Higher gas price = Higher priority

GasAndGasPrice

Use gas limit and then gas price. i.e. Higher gas limit = Lower priority

GasFactorAndGasPrice

Calculate and use priority based on gas and gas price. PRIORITY = GAS_PRICE - GAS/215 * MIN_GAS_PRICE

Rationale: Heavy transactions are paying linear cost (GAS * GAS_PRICE) while the computation might be more expensive.

i.e. 1M gas tx with gas_price=30*min has the same priority as 32k gas tx with gas_price=min