Struct ethcore::miner::pool_client::PoolClient

pub struct PoolClient<'a, C: 'a> { /* fields omitted */ }

Blockchain accesss for transaction pool.

Methods

impl<'a, C: 'a> PoolClient<'a, C> where
C: BlockInfo + CallContract,

pub fn new(
chain: &'a C,
cache: &'a NonceCache,
engine: &'a EthEngine,
accounts: Option<&'a AccountProvider>,
refuse_service_transactions: bool
) -> Self

Creates new client given chain, nonce cache, accounts and service transaction verifier.

pub fn verify_signed(&self, tx: &SignedTransaction) -> Result<(), Error>

Verifies if signed transaction is executable.

This should perform any verifications that rely on chain status.