Struct ethcore::miner::Miner

pub struct Miner { /* fields omitted */ }

Keeps track of transactions using priority queue and holds currently mined block. Handles preparing work for "work sealing" or seals "internally" if Engine does not require work.

Methods

impl Miner

pub fn add_work_listener(&self, notifier: Box<NotifyWork>)

Push listener that will handle new jobs

pub fn add_transactions_listener(&self, f: Box<Fn(&[H256]) + Send + Sync>)

Set a callback to be notified about imported transactions' hashes.

pub fn new(
options: MinerOptions,
gas_pricer: GasPricer,
spec: &Spec,
accounts: Option<Arc<AccountProvider>>
) -> Self

Creates new instance of miner Arc.

pub fn new_for_tests(
spec: &Spec,
accounts: Option<Arc<AccountProvider>>
) -> Miner

Creates new instance of miner with given spec and accounts.

NOTE This should be only used for tests.

pub fn clear(&self)

Clear all pending block states

pub fn update_transaction_queue_limits(&self, block_gas_limit: U256)

Updates transaction queue verification limits.

Limits consist of current block gas limit and minimal gas price.