Struct ethcore::miner::MinerOptions

pub struct MinerOptions {
    pub new_work_notify: Vec<String>,
    pub force_sealing: bool,
    pub reseal_on_external_tx: bool,
    pub reseal_on_own_tx: bool,
    pub reseal_min_period: Duration,
    pub tx_gas_limit: U256,
    pub tx_queue_size: usize,
    pub pending_set: PendingSet,
    pub work_queue_size: usize,
    pub enable_resubmission: bool,
}

Configures the behaviour of the miner.

Fields

new_work_notify

URLs to notify when there is new work.

force_sealing

Force the miner to reseal, even when nobody has asked for work.

reseal_on_external_tx

Reseal on receipt of new external transactions.

reseal_on_own_tx

Reseal on receipt of new local transactions.

reseal_min_period

Minimum period between transaction-inspired reseals.

tx_gas_limit

Maximum amount of gas to bother considering for block insertion.

tx_queue_size

Maximum size of the transaction queue.

pending_set

Whether we should fallback to providing all the queue's transactions or just pending.

work_queue_size

How many historical work packages can we store before running out?

enable_resubmission

Can we submit two different solutions for the same block and expect both to result in an import?