Struct ethcore::miner::MinerStatus

pub struct MinerStatus {
    pub transactions_in_pending_queue: usize,
    pub transactions_in_future_queue: usize,
    pub transactions_in_pending_block: usize,
}

Mining status

Fields
transactions_in_pending_queue: usize

Number of transactions in queue with state pending (ready to be included in block)

transactions_in_future_queue: usize

Number of transactions in queue with state future (not yet ready to be included in block)

transactions_in_pending_block: usize

Number of transactions included in currently mined block