Struct ethcore::miner::Stratum

pub struct Stratum { /* fields omitted */ }

Wrapper for dedicated stratum service

Methods

impl Stratum

fn start(options: &Options,
miner: Weak<Miner>,
client: Weak<Client>)
-> Result<Stratum, Error>

New stratum job dispatcher, given the miner, client and dedicated stratum service

fn register(cfg: &Options,
miner: Arc<Miner>,
client: Weak<Client>)
-> Result<(), Error>

Start STRATUM job dispatcher and register it in the miner