Struct ethcore::miner::TransactionQueue

pub struct TransactionQueue {
    // some fields omitted
}

TransactionQueue implementation

Methods

impl TransactionQueue

fn new() -> Self

Creates new instance of this Queue

fn with_limit(limit: usize) -> Self

Create new instance of this Queue with specified limits

fn set_limit(&mut self, limit: usize)

Set the new limit for current and future queue.

fn limit(&self) -> usize

Returns current limit of transactions in the queue.

fn minimal_gas_price(&self) -> &U256

Get the minimal gas price.

fn set_minimal_gas_price(&mut self, min_gas_price: U256)

Sets new gas price threshold for incoming transactions. Any transaction already imported to the queue is not affected.

fn set_gas_limit(&mut self, gas_limit: U256)

Sets new gas limit. Transactions with gas slightly (GAS_LIMIT_HYSTERESIS) above the limit won't be imported. Any transaction already imported to the queue is not affected.

fn status(&self) -> TransactionQueueStatus

Returns current status for this queue

fn add<T>(&mut self, tx: SignedTransaction, fetch_account: &T, origin: TransactionOrigin) -> Result<TransactionImportResult, Error> where T: Fn(&Address) -> AccountDetails

Add signed transaction to queue to be verified and imported

fn remove_all(&mut self, sender: Address, client_nonce: U256)

Removes all transactions from particular sender up to (excluding) given client (state) nonce. Client (State) Nonce = next valid nonce for this sender.

fn remove_invalid<T>(&mut self, transaction_hash: &H256, fetch_account: &T) where T: Fn(&Address) -> AccountDetails

Removes invalid transaction identified by hash from queue. Assumption is that this transaction nonce is not related to client nonce, so transactions left in queue are processed according to client nonce.

If gap is introduced marks subsequent transactions as future

fn top_transactions(&self) -> Vec<SignedTransaction>

Returns top transactions from the queue ordered by priority.

fn pending_hashes(&self) -> Vec<H256>

Returns hashes of all transactions from current, ordered by priority.

fn has_local_pending_transactions(&self) -> bool

Returns true if there is at least one local transaction pending

fn find(&self, hash: &H256) -> Option<SignedTransaction>

Finds transaction in the queue by hash (if any)

fn clear(&mut self)

Removes all elements (in any state) from the queue

fn last_nonce(&self, address: &Address) -> Option<U256>

Returns highest transaction nonce for given address.