Enum ethcore::service::SyncMessage

pub enum SyncMessage {
    NewChainBlocks {
        imported: Vec<H256>,
        invalid: Vec<H256>,
        retracted: Vec<H256>,
        enacted: Vec<H256>,
    },
    NewChainHead,
    BlockVerified,
    NewTransactions(Vec<Bytes>),
    StartNetwork,
    StopNetwork,
}

Message type for external and internal events

Variants

NewChainBlocks

New block has been imported into the blockchain

Fields

imported

Hashes of blocks imported to blockchain

invalid

Hashes of blocks not imported to blockchain (because were invalid)

retracted

Hashes of blocks that were removed from canonical chain

enacted

Hashes of blocks that are now included in cannonical chain

NewChainHead

Best Block Hash in chain has been changed

BlockVerified

A block is ready

NewTransactions

New transaction RLPs are ready to be imported

StartNetwork

Start network command.

StopNetwork

Stop network command.