Struct ethcore::service::ClientService

pub struct ClientService { /* fields omitted */ }

Client service setup. Creates and registers client and network services with the IO subsystem.

Methods

impl ClientService

fn start(config: ClientConfig,
spec: &Spec,
client_path: &Path,
snapshot_path: &Path,
ipc_path: &Path,
miner: Arc<Miner>)
-> Result<ClientService, Error>

Start the ClientService.

fn add_node(&mut self, _enode: &str)

Add a node to network

fn register_io_handler(&self,
handler: Arc<IoHandler<ClientIoMessage> + Send>)
-> Result<(), IoError>

Get general IO interface

fn client(&self) -> Arc<Client>

Get client interface

fn snapshot_service(&self) -> Arc<SnapshotService>

Get snapshot interface.

fn io(&self) -> Arc<IoService<ClientIoMessage>>

Get network service component

fn add_notify(&self, notify: Arc<ChainNotify>)

Set the actor to be notified on certain chain events