Function ethcore::snapshot::chunk_secondary

pub fn chunk_secondary<'a>(
    chunker: Box<SnapshotComponents>, 
    chain: &'a BlockChain, 
    start_hash: H256, 
    writer: &Mutex<SnapshotWriter + 'a>, 
    progress: &'a Progress
) -> Result<Vec<H256>, Error>

Create and write out all secondary chunks to disk, returning a vector of all the hashes of secondary chunks created.

Secondary chunks are engine-specific, but they intend to corroborate the state data in the state chunks. Returns a list of chunk hashes, with the first having the blocks furthest from the genesis.