Module ethcore::snapshot::io

Snapshot i/o. Ways of writing and reading snapshots. This module supports writing and reading snapshots of two different formats: packed and loose. Packed snapshots are written to a single file, and loose snapshots are written to multiple files in one directory.

Structs

LooseReader
reader for "loose" snapshots

LooseWriter
A "loose" writer writes chunk files into a directory.

PackedReader
Packed snapshot reader.

PackedWriter
A packed snapshot writer. This writes snapshots to a single concatenated file.

Traits

SnapshotReader
Something which can read compressed snapshots.

SnapshotWriter
Something which can write snapshots. Writing the same chunk multiple times will lead to implementation-defined behavior, and is not advised.