Struct ethcore::snapshot::io::PackedReader

pub struct PackedReader { /* fields omitted */ }

Packed snapshot reader.

Methods

impl PackedReader

fn new(path: &Path) -> Result<Option<Self>, Error>

Create a new PackedReader for the file at the given path. This will fail if any io errors are encountered or the file is not a valid packed snapshot.