Struct ethcore::snapshot::io::PackedWriter

pub struct PackedWriter { /* fields omitted */ }

A packed snapshot writer. This writes snapshots to a single concatenated file.

The file format is very simple and consists of three parts: [Concatenated chunk data] [manifest as RLP] [manifest start offset (8 bytes little-endian)]

The manifest contains all the same information as a standard ManifestData, but also maps chunk hashes to their lengths and offsets in the file for easy reading.

Methods

impl PackedWriter

fn new(path: &Path) -> Result<Self>

Create a new "PackedWriter", to write into the file at the given path.