Trait ethcore::snapshot::io::SnapshotWriter

pub trait SnapshotWriter {
    fn write_state_chunk(&mut self, hash: H256, chunk: &[u8]) -> Result<()>;

    fn write_block_chunk(&mut self, hash: H256, chunk: &[u8]) -> Result<()>;

    fn finish(self, manifest: ManifestData) -> Result<()>
    where
        Self: Sized;
}

Something which can write snapshots. Writing the same chunk multiple times will lead to implementation-defined behavior, and is not advised.

Required methods

fn write_state_chunk(&mut self, hash: H256, chunk: &[u8]) -> Result<()>

Write a compressed state chunk.

fn write_block_chunk(&mut self, hash: H256, chunk: &[u8]) -> Result<()>

Write a compressed block chunk.

fn finish(self, manifest: ManifestData) -> Result<()> where
Self: Sized,

Complete writing. The manifest’s chunk lists must be consistent with the chunks written.