Trait ethcore::snapshot::service::SnapshotService

pub trait SnapshotService {
    fn manifest(&self) -> Option<ManifestData>;
    fn chunk(&self, hash: H256) -> Option<Bytes>;
    fn status(&self) -> RestorationStatus;
    fn chunks_done(&self) -> (usize, usize);
    fn begin_restore(&self, manifest: ManifestData);
    fn abort_restore(&self);
    fn restore_state_chunk(&self, hash: H256, chunk: Bytes);
    fn restore_block_chunk(&self, hash: H256, chunk: Bytes);
}

The interface for a snapshot network service. This handles: - restoration of snapshots to temporary databases. - responding to queries for snapshot manifests and chunks

Required Methods

fn manifest(&self) -> Option<ManifestData>

Query the most recent manifest data.

fn chunk(&self, hash: H256) -> Option<Bytes>

Get raw chunk for a given hash.

fn status(&self) -> RestorationStatus

Ask the snapshot service for the restoration status.

fn chunks_done(&self) -> (usize, usize)

Ask the snapshot service for the number of chunks completed. Return a tuple of (state_chunks, block_chunks). Undefined when not restoring.

fn begin_restore(&self, manifest: ManifestData)

Begin snapshot restoration. If restoration in-progress, this will reset it. From this point on, any previous snapshot may become unavailable.

fn abort_restore(&self)

Abort an in-progress restoration if there is one.

fn restore_state_chunk(&self, hash: H256, chunk: Bytes)

Feed a raw state chunk to the service to be processed asynchronously. no-op if not currently restoring.

fn restore_block_chunk(&self, hash: H256, chunk: Bytes)

Feed a raw block chunk to the service to be processed asynchronously. no-op if currently restoring.