Struct ethcore::snapshot::PoaSnapshot

pub struct PoaSnapshot;

Snapshot creation and restoration for PoA chains. Chunk format:

[FLAG, [header, epoch data], ...]

Header data at which transition occurred,

epoch data (usually list of validators and proof of change)

FLAG is a bool: true for last chunk, false otherwise.

The last item of the last chunk will be a list containing data for the warp target block: [header, transactions, uncles, receipts, parent_td].