Struct ethcore::snapshot::PowSnapshot

pub struct PowSnapshot {
    pub blocks: u64,
    pub max_restore_blocks: u64,
}

Snapshot creation and restoration for PoW chains. This includes blocks from the head of the chain as a loose assurance that the chain is valid.

Fields
blocks: u64

Number of blocks from the head of the chain to include in the snapshot.

max_restore_blocks: u64

Number of to allow in the snapshot when restoring.

Methods

impl PowSnapshot

pub fn new(blocks: u64, max_restore_blocks: u64) -> PowSnapshot

Create a new instance.