Struct ethcore::snapshot::StateRebuilder

pub struct StateRebuilder { /* fields omitted */ }

Used to rebuild the state trie piece by piece.

Methods

impl StateRebuilder

fn new(db: Arc<Database>, pruning: Algorithm) -> Self

Create a new state rebuilder to write into the given backing DB.

fn feed(&mut self, chunk: &[u8], flag: &AtomicBool) -> Result<(), Error>

Feed an uncompressed state chunk into the rebuilder.

fn check_missing(self) -> Result<(), Error>

Check for accounts missing code. Once all chunks have been fed, there should be none.

fn state_root(&self) -> H256

Get the state root of the rebuilder.