Struct ethcore::snapshot::Watcher

pub struct Watcher { /* fields omitted */ }

A ChainNotify implementation which will trigger a snapshot event at certain block numbers.

Methods

impl Watcher

fn new<F>(client: Arc<Client>,
sync_status: F,
channel: IoChannel<ClientIoMessage>,
period: u64,
history: u64)
-> Self where F: 'static + Send + Sync + Fn() -> bool

Create a new Watcher which will trigger a snapshot event once every period blocks, but only after that block is history blocks old.