Module ethcore::spec

Blockchain params.

Structs

CommonParams
Parameters common to ethereum-like blockchains. NOTE: when adding bugfix hard-fork parameters, add to nonzero_bugfix_hard_fork

Genesis
Genesis components.

Spec
Parameters for a block chain; includes both those intrinsic to the design of the chain and those to be interpreted by the active chain engine.

SpecHardcodedSync
Part of Spec. Describes the hardcoded synchronization parameters.

SpecParams
Runtime parameters for the spec that are related to how the software should run the chain, rather than integral properties of the chain itself.

Enums

OptimizeFor