Struct ethcore::spec::spec::Spec

pub struct Spec {
    pub name: String,
    pub engine: Box<Engine>,
    pub nodes: Vec<String>,
    pub params: CommonParams,
    pub parent_hash: H256,
    pub author: Address,
    pub difficulty: U256,
    pub gas_limit: U256,
    pub gas_used: U256,
    pub timestamp: u64,
    pub transactions_root: H256,
    pub receipts_root: H256,
    pub extra_data: Bytes,
    pub seal_fields: usize,
    pub seal_rlp: Bytes,
    // some fields omitted
}

Parameters for a block chain; includes both those intrinsic to the design of the chain and those to be interpreted by the active chain engine.

Fields

name

User friendly spec name

engine

What engine are we using for this?

nodes

Known nodes on the network in enode format.

params

Parameters common to all engines.

parent_hash

The genesis block's parent hash field.

author

The genesis block's author field.

difficulty

The genesis block's difficulty field.

gas_limit

The genesis block's gas limit field.

gas_used

The genesis block's gas used field.

timestamp

The genesis block's timestamp field.

transactions_root

Transactions root of the genesis block. Should be SHA3_NULL_RLP.

receipts_root

Receipts root of the genesis block. Should be SHA3_NULL_RLP.

extra_data

The genesis block's extra data field.

seal_fields

The number of seal fields in the genesis block.

seal_rlp

Each seal field, expressed as RLP, concatenated.

Methods

impl Spec

fn state_root(&self) -> H256

Return the state root for the genesis state, memoising accordingly.

fn nodes(&self) -> &Vec<String>

Get the known knodes of the network in enode format.

fn network_id(&self) -> U256

Get the configured Network ID.

fn genesis_header(&self) -> Header

Get the header of the genesis block.

fn genesis_block(&self) -> Bytes

Compose the genesis block for this chain.

fn overwrite_genesis_params(&mut self, g: Genesis)

Overwrite the genesis components.

fn set_genesis_state(&mut self, s: PodState)

Alter the value of the genesis state.

fn is_state_root_valid(&self) -> bool

Returns false if the memoized state root is invalid. true otherwise.

fn ensure_db_good(&self, db: &mut HashDB) -> bool

Ensure that the given state DB has the trie nodes in for the genesis state.

fn load(reader: &[u8]) -> Self

Loads spec from json file.

fn new_test() -> Spec

Create a new Spec which conforms to the Frontier-era Morden chain except that it's a NullEngine consensus.

fn new_null() -> Spec

Create a new Spec which is a NullEngine consensus with a premine of address whose secret is sha3('').