Struct ethcore::spec::Spec

pub struct Spec {
    pub name: String,
    pub engine_name: String,
    pub engine_params: HashMap<String, Bytes>,
    pub builtins: HashMap<Address, Builtin>,
    pub parent_hash: H256,
    pub author: Address,
    pub difficulty: U256,
    pub gas_limit: U256,
    pub gas_used: U256,
    pub timestamp: u64,
    pub extra_data: Bytes,
    pub genesis_state: HashMap<Address, GenesisAccount>,
    pub seal_fields: usize,
    pub seal_rlp: Bytes,
    // some fields omitted
}

Parameters for a block chain; includes both those intrinsic to the design of the chain and those to be interpreted by the active chain engine.

Fields

name

TODO [Gav Wood] Please document me

engine_name

TODO [Gav Wood] Please document me

engine_params

TODO [Gav Wood] Please document me

builtins

TODO [Gav Wood] Please document me

parent_hash

TODO [Gav Wood] Please document me

author

TODO [Gav Wood] Please document me

difficulty

TODO [Gav Wood] Please document me

gas_limit

TODO [Gav Wood] Please document me

gas_used

TODO [Gav Wood] Please document me

timestamp

TODO [Gav Wood] Please document me

extra_data

TODO [arkpar] Please document me

genesis_state

TODO [Gav Wood] Please document me

seal_fields

TODO [Gav Wood] Please document me

seal_rlp

TODO [Gav Wood] Please document me

Methods

impl Spec

fn to_engine(self) -> Result<Box<Engine>, Error>

Convert this object into a boxed Engine of the right underlying type.

fn state_root(&self) -> H256

Return the state root for the genesis state, memoising accordingly.

fn genesis_header(&self) -> Header

TODO [Gav Wood] Please document me

fn genesis_block(&self) -> Bytes

Compose the genesis block for this chain.

impl Spec

fn ensure_db_good(&self, db: &mut HashDB) -> bool

Ensure that the given state DB has the trie nodes in for the genesis state.

fn from_json_utf8(data: &[u8]) -> Spec

Create a new Spec from a JSON UTF-8 data resource data.

fn from_json_str(s: &str) -> Spec

Create a new Spec from a JSON string.

fn new_test() -> Spec

Create a new Spec which conforms to the Morden chain except that it's a NullEngine consensus.