Module ethcore::state::backend

A minimal "state backend" trait: an abstraction over the sources of data a blockchain state may draw upon.

Currently assumes a very specific DB + cache structure, but should become general over time to the point where not even a merkle trie is strictly necessary.

Structs

Basic
A basic backend. Just wraps the given database, directly inserting into and deleting from it. Doesn't cache anything.

ProofCheck
A raw backend used to check proofs of execution.

Proving
Proving state backend. This keeps track of all state values loaded during usage of this backend. The proof-of-execution can be extracted with extract_proof.

Traits

Backend
State backend. See module docs for more details.