Struct ethcore::state::backend::Proving

pub struct Proving<H: AsHashDB> { /* fields omitted */ }

Proving state backend. This keeps track of all state values loaded during usage of this backend. The proof-of-execution can be extracted with extract_proof.

This doesn't cache anything or rely on the canonical state caches.

Methods

impl<H: AsHashDB> Proving<H>

pub fn new(base: H) -> Self

Create a new Proving over a base database. This will store all values ever fetched from that base.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn extract_proof(self) -> Vec<DBValue>

Consume the backend, extracting the gathered proof in lexicographical order by value.