Struct ethcore::state::ApplyOutcome

pub struct ApplyOutcome<T, V> {
    pub receipt: Receipt,
    pub output: Bytes,
    pub trace: Vec<T>,
    pub vm_trace: Option<V>,
}

Used to return information about an State::apply operation.

Fields
receipt: Receipt

The receipt for the applied transaction.

output: Bytes

The output of the applied transaction.

trace: Vec<T>

The trace for the applied transaction, empty if tracing was not produced.

vm_trace: Option<V>

The VM trace for the applied transaction, None if tracing was not produced.