Module ethcore::test_helpers

Set of different helpers for client tests

Structs

TestNotify
Test actor for chain events

Functions

create_test_block
Creates test block with corresponding header

create_test_block_with_data
Creates test block with corresponding header and data

generate_dummy_blockchain
Generates dummy blockchain with corresponding amount of blocks

generate_dummy_blockchain_with_extra
Generates dummy blockchain with corresponding amount of blocks (using creation with extra method for blocks creation)

generate_dummy_client
Generates dummy client (not test client) with corresponding amount of blocks

generate_dummy_client_with_data
Generates dummy client (not test client) with corresponding amount of blocks and txs per every block

generate_dummy_client_with_spec_accounts_and_data
Generates dummy client (not test client) with corresponding blocks, accounts and spec

generate_dummy_client_with_spec_and_accounts
Generates dummy client (not test client) with corresponding spec and accounts

generate_dummy_client_with_spec_and_data
Generates dummy client (not test client) with corresponding amount of blocks, txs per block and spec

generate_dummy_empty_blockchain
Returns empty dummy blockchain

get_bad_state_dummy_block
Returns hash of the dummy block with incorrect state root

get_good_dummy_block
Returns hash of the correct dummy block

get_good_dummy_block_fork_seq
Returns sequence of hashes of the dummy blocks beginning from corresponding parent

get_good_dummy_block_hash
Returns hash and header of the correct dummy block

get_good_dummy_block_seq
Returns sequence of hashes of the dummy blocks

get_temp_state
Returns temp state

get_temp_state_db
Returns temp state db

get_temp_state_with_factory
Returns temp state using coresponding factory

get_test_client_with_blocks
Creates dummy client (not test client) with corresponding blocks

new_db
Creates new test instance of BlockChainDB

push_block_with_transactions
Adds one block with transactions

push_blocks_to_client
Adds blocks to the client

restoration_db_handler
Creates new instance of KeyValueDBHandler