Enum ethcore::trace::TraceError

pub enum TraceError {
    OutOfGas,
    BadJumpDestination,
    BadInstruction,
    StackUnderflow,
    OutOfStack,
    BuiltIn,
    Internal,
    MutableCallInStaticContext,
    Wasm,
    OutOfBounds,
    Reverted,
}

Trace evm errors.

Variants
OutOfGas

OutOfGas is returned when transaction execution runs out of gas.

BadJumpDestination

BadJumpDestination is returned when execution tried to move to position that wasn't marked with JUMPDEST instruction

BadInstruction

BadInstructions is returned when given instruction is not supported

StackUnderflow

StackUnderflow when there is not enough stack elements to execute instruction

OutOfStack

When execution would exceed defined Stack Limit

BuiltIn

When builtin contract failed on input data

Internal

Returned on evm internal error. Should never be ignored during development. Likely to cause consensus issues.

MutableCallInStaticContext

When execution tries to modify the state in static context

Wasm

Wasm error

OutOfBounds

Contract tried to access past the return data buffer.

Reverted

Execution has been reverted with REVERT instruction.