Struct ethcore::trace::filter::AddressesFilter

pub struct AddressesFilter {
    // some fields omitted
}

Addresses filter.

Used to create bloom possibilities and match filters.

Methods

impl AddressesFilter

fn matches(&self, address: &Address) -> bool

Returns true if address matches one of the searched addresses.

fn matches_all(&self) -> bool

Returns true if this address filter matches everything.

fn blooms(&self) -> Vec<LogBloom>

Returns blooms of this addresses filter.

fn with_blooms(&self, blooms: Vec<LogBloom>) -> Vec<LogBloom>

Returns vector of blooms zipped with blooms of this addresses filter.