Struct ethcore::trace::flat::FlatTrace

pub struct FlatTrace {
    pub action: Action,
    pub result: Res,
    pub subtraces: usize,
    pub trace_address: VecDeque<usize>,
}

Trace localized in vector of traces produced by a single transaction.

Parent and children indexes refer to positions in this vector.

Fields
action: Action

Type of action performed by a transaction.

result: Res

Result of this action.

subtraces: usize

Number of subtraces.

trace_address: VecDeque<usize>

Exact location of trace.

[index in root, index in first CALL, index in second CALL, ...]

Methods

impl FlatTrace

fn bloom(&self) -> H2048

Returns bloom of the trace.