Struct ethcore::trace::Trace

pub struct Trace {
    pub depth: usize,
    pub action: Action,
    pub subs: Vec<Trace>,
    pub result: Res,
}

A trace; includes a description of the action being traced and sub traces of each interior action.

Fields

depth

The number of EVM execution environments active when this action happened; 0 if it's the outer action of the transaction.

action

The action being performed.

subs

The sub traces for each interior action performed as part of this call.

result

The result of the performed action.