Struct ethcore::trace::TraceCreate

pub struct TraceCreate {
    pub from: Address,
    pub value: U256,
    pub gas: U256,
    pub init: Bytes,
    pub result: Option<(U256, Address, Bytes)>,
}

Description of a create action, either a CREATE operation or a create transction.

Fields

from

The address of the creator.

value

The value with which the new account is endowed.

gas

The gas available for the creation init code.

init

The init code.

result

The result of the operation; tuple of the gas used, the address of the newly created account and its code. NOTE: Presently failed operations are not reported so this will always be Some.