Trait ethcore::trace::DatabaseExtras

pub trait DatabaseExtras {
    fn block_hash(&self, block_number: BlockNumber) -> Option<H256>;

    fn transaction_hash(
        &self, 
        block_number: BlockNumber, 
        tx_position: usize
    ) -> Option<H256>;
}

DbExtras provides an interface to query extra data which is not stored in tracesdb, but necessary to work correctly.

Required Methods

fn block_hash(&self, block_number: BlockNumber) -> Option<H256>

Returns hash of given block number.

fn transaction_hash(
&self,
block_number: BlockNumber,
tx_position: usize
) -> Option<H256>

Returns hash of transaction at given position.