Struct ethcore::tree_route::TreeRoute

pub struct TreeRoute {
    pub blocks: Vec<H256>,
    pub ancestor: H256,
    pub index: usize,
}

Represents a tree route between from block and to block:

Fields
blocks: Vec<H256>

A vector of hashes of all blocks, ordered from from to to.

ancestor: H256

Best common ancestor of these blocks.

index: usize

An index where best common ancestor would be.