Module ethcore::verification::queue

A queue of blocks. Sits between network or other I/O and the BlockChain. Sorts them ready for blockchain insertion.

Modules

kind

Definition of valid items for the verification queue.

Structs

Config

Verification queue configuration

QueueInfo

Verification queue status

VerificationQueue

A queue of items to be verified. Sits between network or other I/O and the BlockChain. Keeps them in the same order as inserted, minus invalid items.

VerifierSettings

Verifier settings.

Verifying

An item which is in the process of being verified.

Enums

Status

Status of items in the queue.

Type Definitions

BlockQueue

Type alias for block queue convenience.

HeaderQueue

Type alias for header queue convenience.