Struct ethcore::verification::queue::Config

pub struct Config {
    pub max_queue_size: usize,
    pub max_mem_use: usize,
    pub verifier_settings: VerifierSettings,
}

Verification queue configuration

Fields
max_queue_size: usize

Maximum number of items to keep in unverified queue. When the limit is reached, is_full returns true.

max_mem_use: usize

Maximum heap memory to use. When the limit is reached, is_full returns true.

verifier_settings: VerifierSettings

Settings for the number of verifiers and adaptation strategy.