Trait ethcore::verification::Verifier

pub trait Verifier<C>: Send + Sync where
    C: BlockInfo + CallContract, {
    fn verify_block_family(
        &self, 
        header: &Header, 
        parent: &Header, 
        engine: &dyn EthEngine, 
        do_full: Option<FullFamilyParams<'_, C>>
    ) -> Result<(), Error>;

    fn verify_block_final(
        &self, 
        expected: &Header, 
        got: &Header
    ) -> Result<(), Error>;

    fn verify_block_external(
        &self, 
        header: &Header, 
        engine: &dyn EthEngine
    ) -> Result<(), Error>;
}

Should be used to verify blocks.

Required methods

fn verify_block_family(
&self,
header: &Header,
parent: &Header,
engine: &dyn EthEngine,
do_full: Option<FullFamilyParams<'_, C>>
) -> Result<(), Error>

Verify a block relative to its parent and uncles.

fn verify_block_final(
&self,
expected: &Header,
got: &Header
) -> Result<(), Error>

Do a final verification check for an enacted header vs its expected counterpart.

fn verify_block_external(
&self,
header: &Header,
engine: &dyn EthEngine
) -> Result<(), Error>

Verify a block, inspecing external state.