Module ethcore::verification::verification

Block and transaction verification functions

Block verification is done in 3 steps

Quick verification upon adding to the block queue

Signatures verification done in the queue.

Final verification against the blockchain done before enactment.

Structs

FullFamilyParams
Parameters for full verification of block family

PreverifiedBlock
Preprocessed block data gathered in verify_block_unordered call

Functions

verify_block_basic
Phase 1 quick block verification. Only does checks that are cheap. Operates on a single block

verify_block_family
Phase 3 verification. Check block information against parent and uncles.

verify_block_final
Phase 4 verification. Check block information against transaction enactment results,

verify_block_unordered
Phase 2 verification. Perform costly checks such as transaction signatures and block nonce for ethash. Still operates on a individual block Returns a PreverifiedBlock structure populated with transactions

verify_header_params
Check basic header parameters.