Crate ethcore

Ethcore library

Re-exports

pub use executive::contract_address;

Modules

block

Base data structure of this module is Block.

client

Blockchain database client.

engines

Consensus engine specification and basic implementations.

error

General error types for use in ethcore.

ethereum

Ethereum protocol module.

executed

Transaction execution format module.

executive

Transaction Execution environment.

machine

Generalization of a state machine for a consensus engine.

miner

Miner module Keeps track of transactions and currently sealed pending block.

pod_account

Account system expressed in Plain Old Data.

pod_state

State of all accounts in the system expressed in Plain Old Data.

snapshot

Snapshot creation, restoration, and network service.

spec

Blockchain params.

state

A mutable state representation suitable to execute transactions. Generic over a Backend. Deals with Accounts. Unconfirmed sub-states are managed with checkpoints which may be canonicalized or rolled back.

state_db

State database abstraction. For more info, see the doc for StateDB

trace

Tracing

transaction_ext

Ethereum transaction

verification

Block verification utilities.

Enums

CreateContractAddress

Specifies how an address is calculated for a new contract.

TrieSpec

Trie types