Crate ethcore

Ethcore library

Rust version:

nightly

Supported platforms:

OSX

Linux

Building:

Ubuntu 14.04 and later:

# install multirust
curl -sf https://raw.githubusercontent.com/brson/multirust/master/blastoff.sh | sh -s -- --yes

# export rust LIBRARY_PATH
export LIBRARY_PATH=/usr/local/lib

# download and build parity
git clone https://github.com/ethcore/parity
cd parity
multirust override beta
cargo build --release

OSX:

# install rocksdb && multirust
brew update
brew install multirust

# export rust LIBRARY_PATH
export LIBRARY_PATH=/usr/local/lib

# download and build parity
git clone https://github.com/ethcore/parity
cd parity
multirust override beta
cargo build --release

Reexports

pub extern crate ethstore;

Modules

account_diff
Diff between two accounts.

account_provider
Account management.

action_params
Evm input params.

block
Blockchain block.

block_import_error
Block import error related types

block_status
Block status description module

blockchain_info
Blockhain info type definition

call_analytics
Call analytics related types

client
Blockchain database client.

db
Database utilities and definitions.

encoded
Lazily-decoded owning views of RLP-encoded blockchain objects. These views are meant to contain trusted data -- without encoding errors or inconsistencies.

engines
Consensus engine specification and basic implementations.

error
General error types for use in ethcore.

ethereum
Ethereum protocol module.

evm
Ethereum virtual machine.

executed
Transaction execution format module.

filter
Blockchain filter

header
Block header.

ids
Unique identifiers.

log_entry
Log entry type definition.

migrations
Database migrations.

miner
Miner module Keeps track of transactions and mined block.

mode
Mode type

pod_state
State of all accounts in the system expressed in Plain Old Data.

pruning_info
Information about portions of the state and chain which the client may serve.

receipt
Receipt

restoration_status
Restoration status type definition

security_level
Indication of how secure the chain is.

service
Creates and registers client and network services.

snapshot
Snapshot creation, restoration, and network service.

snapshot_manifest
Snapshot manifest type definition

spec
Blockchain params.

state_diff
State diff module.

trace
Tracing

trace_filter
Trace filter related types

trace_types
Types used in the public api

transaction
Transaction data structure.

transaction_import
Transaction import result related types

tree_route
Tree route info type definition

verification
Block verification utilities.

verification_queue_info
Verification queue info types

views
Block oriented views onto rlp.

Macros

evm_test
Create tests by injecting different VM factories

evm_test_ignore
Create ignored tests by injecting different VM factories

Functions

contract_address
Returns new address created from address and given nonce.