Struct ethcore::block::ClosedBlock

pub struct ClosedBlock<'x, 'y> {
    // some fields omitted
}

Just like OpenBlock, except that we've applied Engine::on_close_block, finished up the non-seal header fields, and collected the uncles.

There is no function available to push a transaction. If you want that you'll need to reopen() it.

Methods

impl<'x, 'y> ClosedBlock<'x, 'y>

fn hash(&self) -> H256

Get the hash of the header without seal arguments.

fn seal(self, seal: Vec<Bytes>) -> Result<SealedBlock, BlockError>

Provide a valid seal in order to turn this into a SealedBlock.

NOTE: This does not check the validity of seal with the engine.

fn reopen(self) -> OpenBlock<'x, 'y>

Turn this back into an OpenBlock.

fn drain(self) -> JournalDB

Drop this object and return the underlieing database.