Struct ethcore::block::ClosedBlock

pub struct ClosedBlock { /* fields omitted */ }

Just like OpenBlock, except that we've applied Engine::on_close_block, finished up the non-seal header fields, and collected the uncles.

There is no function available to push a transaction.

Methods

impl ClosedBlock

fn hash(&self) -> H256

Get the hash of the header without seal arguments.

fn lock(self) -> LockedBlock

Turn this into a LockedBlock, unable to be reopened again.

fn reopen(self, engine: &Engine) -> OpenBlock

Given an engine reference, reopen the ClosedBlock into an OpenBlock.