Struct ethcore::block::OpenBlock

pub struct OpenBlock<'x, 'y> {
    // some fields omitted
}

Block that is ready for transactions to be added.

It's a bit like a Vec, eccept that whenever a transaction is pushed, we execute it and maintain the system state(). We also archive execution receipts in preparation for later block creation.

Methods

impl<'x, 'y> OpenBlock<'x, 'y>

fn new<'a, 'b>(engine: &'a Engine, db: JournalDB, parent: &Header, last_hashes: &'b LastHashes, author: Address, extra_data: Bytes) -> OpenBlock<'a, 'b>

Create a new OpenBlock ready for transaction pushing.

fn set_author(&mut self, author: Address)

Alter the author for the block.

fn set_timestamp(&mut self, timestamp: u64)

Alter the timestamp of the block.

fn set_difficulty(&mut self, a: U256)

Alter the difficulty for the block.

fn set_gas_limit(&mut self, a: U256)

Alter the gas limit for the block.

fn set_gas_used(&mut self, a: U256)

Alter the gas limit for the block.

fn set_extra_data(&mut self, extra_data: Bytes) -> Result<(), BlockError>

Alter the extra_data for the block.

fn push_uncle(&mut self, valid_uncle_header: Header) -> Result<(), BlockError>

Add an uncle to the block, if possible.

NOTE Will check chain constraints and the uncle number but will NOT check that the header itself is actually valid.

fn env_info(&self) -> EnvInfo

Get the environment info concerning this block.

fn push_transaction(&mut self, t: Transaction, h: Option<H256>) -> Result<&Receipt, Error>

Push a transaction into the block.

If valid, it will be executed, and archived together with the receipt.

fn close(self) -> ClosedBlock<'x, 'y>

Turn this into a ClosedBlock. A BlockChain must be provided in order to figure out the uncles.