Struct ethcore::block::OpenBlock

pub struct OpenBlock<'x> { /* fields omitted */ }

Block that is ready for transactions to be added.

It's a bit like a Vec, except that whenever a transaction is pushed, we execute it and maintain the system state(). We also archive execution receipts in preparation for later block creation.

Methods

impl<'x> OpenBlock<'x>

pub fn new<'a>(
engine: &'x EthEngine,
factories: Factories,
tracing: bool,
db: StateDB,
parent: &Header,
last_hashes: Arc<LastHashes>,
author: Address,
gas_range_target: (U256, U256),
extra_data: Bytes,
is_epoch_begin: bool,
ancestry: &mut Iterator<Item = ExtendedHeader>
) -> Result<Self, Error>

Create a new OpenBlock ready for transaction pushing.

pub fn set_timestamp(&mut self, timestamp: u64)

Alter the timestamp of the block.

pub fn remove_gas_limit(&mut self)

Removes block gas limit.

pub fn push_uncle(
&mut self,
valid_uncle_header: Header
) -> Result<(), BlockError>

Add an uncle to the block, if possible.

NOTE Will check chain constraints and the uncle number but will NOT check that the header itself is actually valid.

pub fn env_info(&self) -> EnvInfo

Get the environment info concerning this block.

pub fn push_transaction(
&mut self,
t: SignedTransaction,
h: Option<H256>
) -> Result<&Receipt, Error>

Push a transaction into the block.

If valid, it will be executed, and archived together with the receipt.

pub fn close(self) -> ClosedBlock

Turn this into a ClosedBlock.

pub fn close_and_lock(self) -> LockedBlock

Turn this into a LockedBlock.