Struct ethcore::block::OpenBlock

pub struct OpenBlock<'x> { /* fields omitted */ }

Block that is ready for transactions to be added.

It's a bit like a Vec, except that whenever a transaction is pushed, we execute it and maintain the system state(). We also archive execution receipts in preparation for later block creation.

Methods

impl<'x> OpenBlock<'x>

fn new(engine: &'x Engine,
factories: Factories,
tracing: bool,
db: StateDB,
parent: &Header,
last_hashes: Arc<LastHashes>,
author: Address,
gas_range_target: (U256, U256),
extra_data: Bytes)
-> Result<Self, Error>

Create a new OpenBlock ready for transaction pushing.

fn set_author(&mut self, author: Address)

Alter the author for the block.

fn set_timestamp(&mut self, timestamp: u64)

Alter the timestamp of the block.

fn set_difficulty(&mut self, a: U256)

Alter the difficulty for the block.

fn set_gas_limit(&mut self, a: U256)

Alter the gas limit for the block.

fn set_gas_used(&mut self, a: U256)

Alter the gas limit for the block.

fn set_uncles_hash(&mut self, h: H256)

Alter the uncles hash the block.

fn set_transactions_root(&mut self, h: H256)

Alter transactions root for the block.

fn set_receipts_root(&mut self, h: H256)

Alter the receipts root for the block.

fn set_extra_data(&mut self, extra_data: Bytes) -> Result<(), BlockError>

Alter the extra_data for the block.

fn push_uncle(&mut self, valid_uncle_header: Header) -> Result<(), BlockError>

Add an uncle to the block, if possible.

NOTE Will check chain constraints and the uncle number but will NOT check that the header itself is actually valid.

fn env_info(&self) -> EnvInfo

Get the environment info concerning this block.

fn push_transaction(&mut self,
t: SignedTransaction,
h: Option<H256>)
-> Result<&Receipt, Error>

Push a transaction into the block.

If valid, it will be executed, and archived together with the receipt.

fn close(self) -> ClosedBlock

Turn this into a ClosedBlock.

fn close_and_lock(self) -> LockedBlock

Turn this into a LockedBlock.