Struct ethcore::builtin::Builtin

pub struct Builtin {
    pub cost: Box<Fn(usize) -> U256>,
    pub execute: Box<Fn(&[u8], &mut [u8])>,
}

Definition of a contract whose implementation is built-in.

Fields

cost

The gas cost of running this built-in for the given size of input data.

execute

Run this built-in function with the input being the first argument and the output being placed into the second.

Methods

impl Builtin

fn new(cost: Box<Fn(usize) -> U256>, execute: Box<Fn(&[u8], &mut [u8])>) -> Builtin

Create a new object from components.

fn from_named_linear(name: &str, base_cost: usize, word_cost: usize) -> Option<Builtin>

Create a new object from a builtin-function name with a linear cost associated with input size.

fn cost(&self, s: usize) -> U256

Simple forwarder for cost.

fn execute(&self, input: &[u8], output: &mut [u8])

Simple forwarder for execute.

fn from_json(json: &Json) -> Option<Builtin>

Create a builtin from JSON.

JSON must be of the form { "name": "identity", "linear": {"base": 10, "word": 20} }.