Struct ethcore::client::Client

pub struct Client {
    // some fields omitted
}

Blockchain database client backed by a persistent database. Owns and manages a blockchain and a block queue.

Methods

impl Client

fn new(spec: Spec, path: &Path, message_channel: IoChannel<NetSyncMessage>) -> Result<Client, Error>

Create a new client with given spec and DB path.

fn import_verified_blocks(&mut self)

This is triggered by a message coming from a block queue when the block is ready for insertion

fn cache_info(&self) -> CacheSize

Get info on the cache.

fn report(&self) -> ClientReport

Get the report.

fn tick(&self)

Tick the client.