Struct ethcore::client::Client

pub struct Client { /* fields omitted */ }

Blockchain database client backed by a persistent database. Owns and manages a blockchain and a block queue. Call import_block() to import a block asynchronously; flush_queue() flushes the queue.

Methods

impl Client

pub fn new(
config: ClientConfig,
spec: &Spec,
db: Arc<BlockChainDB>,
miner: Arc<Miner>,
message_channel: IoChannel<ClientIoMessage>
) -> Result<Arc<Client>, Error>

Create a new client with given parameters. The database is assumed to have been initialized with the correct columns.

pub fn keep_alive(&self)

Wakes up client if it's a sleep.

pub fn add_notify(&self, target: Arc<ChainNotify>)

Adds an actor to be notified on certain events

pub fn set_exit_handler<F>(&self, f: F) where
F: Fn(String) + 'static + Send,

Set a closure to call when the client wants to be restarted.

The parameter passed to the callback is the name of the new chain spec to use after the restart.

pub fn engine(&self) -> &EthEngine

Returns engine reference.

pub fn on_user_defaults_change<F>(&self, f: F) where
F: 'static + FnMut(Option<Mode>) + Send,

Register an action to be done if a mode/spec_name change happens.

pub fn flush_queue(&self)

Flush the block import queue.

pub fn latest_env_info(&self) -> EnvInfo

The env info as of the best block.

pub fn env_info(&self, id: BlockId) -> Option<EnvInfo>

The env info as of a given block. returns None if the block unknown.

pub fn import_verified_blocks(&self) -> usize

This is triggered by a message coming from a block queue when the block is ready for insertion

pub fn set_io_channel(&self, io_channel: IoChannel<ClientIoMessage>)

Replace io channel. Useful for testing.

pub fn latest_state(&self) -> State<StateDB>

Get a copy of the best block's state.

pub fn state_at(&self, id: BlockId) -> Option<State<StateDB>>

Attempt to get a copy of a specific block's final state.

This will not fail if given BlockId::Latest. Otherwise, this can fail (but may not) if the DB prunes state or the block is unknown.

pub fn state_at_beginning(&self, id: BlockId) -> Option<State<StateDB>>

Attempt to get a copy of a specific block's beginning state.

This will not fail if given BlockId::Latest. Otherwise, this can fail (but may not) if the DB prunes state.

Important traits for Box<R>
impl<R> Read for Box<R> where
R: Read + ?Sized, impl<W> Write for Box<W> where
W: Write + ?Sized, impl<I> Iterator for Box<I> where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn state(&self) -> Box<StateInfo>

Get a copy of the best block's state.

pub fn blockchain_cache_info(&self) -> BlockChainCacheSize

Get info on the cache.

pub fn report(&self) -> ClientReport

Get the report.

pub fn tick(&self, prevent_sleep: bool)

Tick the client.

pub fn take_snapshot<W: SnapshotWriter + Send>(
&self,
writer: W,
at: BlockId,
p: &Progress
) -> Result<(), EthcoreError>

Take a snapshot at the given block. If the ID given is "latest", this will default to 1000 blocks behind.

pub fn pruning_history(&self) -> u64

Ask the client what the history parameter is.