Struct ethcore::client::Client

pub struct Client<V = CanonVerifier> where V: Verifier {
    // some fields omitted
}

Blockchain database client backed by a persistent database. Owns and manages a blockchain and a block queue. Call import_block() to import a block asynchronously; flush_queue() flushes the queue.

Methods

impl Client<CanonVerifier>

fn new(config: ClientConfig, spec: Spec, path: &Path, message_channel: IoChannel<NetSyncMessage>) -> Arc<Client>

Create a new client with given spec and DB path.

impl<V> Client<V> where V: Verifier

fn new_with_verifier(config: ClientConfig, spec: Spec, path: &Path, message_channel: IoChannel<NetSyncMessage>) -> Arc<Client<V>>

Create a new client with given spec and DB path and custom verifier.

fn flush_queue(&self)

Flush the block import queue.

fn import_verified_blocks(&self, io: &IoChannel<NetSyncMessage>) -> usize

This is triggered by a message coming from a block queue when the block is ready for insertion

fn state(&self) -> State

Get a copy of the best block's state.

fn blockchain_cache_info(&self) -> BlockChainCacheSize

Get info on the cache.

fn report(&self) -> ClientReport

Get the report.

fn tick(&self)

Tick the client.

fn configure_cache(&self, pref_cache_size: usize, max_cache_size: usize)

Set up the cache behaviour.