Struct ethcore::client::Client

pub struct Client {
    // some fields omitted
}

Blockchain database client backed by a persistent database. Owns and manages a blockchain and a block queue. Call import_block() to import a block asynchronously; flush_queue() flushes the queue.

Methods

impl Client

fn new(config: ClientConfig, spec: &Spec, path: &Path, miner: Arc<Miner>, message_channel: IoChannel<ClientIoMessage>, db_config: &DatabaseConfig) -> Result<Arc<Client>, ClientError>

Create a new client with given spec and DB path and custom verifier.

fn add_notify(&self, target: Arc<ChainNotify>)

Adds an actor to be notified on certain events

fn flush_queue(&self)

Flush the block import queue.

fn import_verified_blocks(&self) -> usize

This is triggered by a message coming from a block queue when the block is ready for insertion

fn import_queued_transactions(&self, transactions: &[Bytes]) -> usize

Import transactions from the IO queue

fn state_at(&self, id: BlockID) -> Option<State>

Attempt to get a copy of a specific block's final state.

This will not fail if given BlockID::Latest. Otherwise, this can fail (but may not) if the DB prunes state.

fn state_at_beginning(&self, id: BlockID) -> Option<State>

Attempt to get a copy of a specific block's beginning state.

This will not fail if given BlockID::Latest. Otherwise, this can fail (but may not) if the DB prunes state.

fn state(&self) -> State

Get a copy of the best block's state.

fn blockchain_cache_info(&self) -> BlockChainCacheSize

Get info on the cache.

fn report(&self) -> ClientReport

Get the report.

fn tick(&self)

Tick the client.

fn block_number(&self, id: BlockID) -> Option<BlockNumber>

Look up the block number for the given block ID.

fn take_snapshot<W: SnapshotWriter + Send>(&self, writer: W, at: BlockID, p: &Progress) -> Result<(), EthcoreError>

Take a snapshot at the given block. If the ID given is "latest", this will default to 1000 blocks behind.

fn pruning_history(&self) -> u64

Ask the client what the history parameter is.