Struct ethcore::client::TestBlockChainClient

pub struct TestBlockChainClient {
    pub blocks: RwLock<HashMap<H256, Bytes>>,
    pub numbers: RwLock<HashMap<usize, H256>>,
    pub genesis_hash: H256,
    pub last_hash: RwLock<H256>,
    pub extra_data: Bytes,
    pub difficulty: RwLock<U256>,
    pub balances: RwLock<HashMap<Address, U256>>,
    pub nonces: RwLock<HashMap<Address, U256>>,
    pub storage: RwLock<HashMap<(Address, H256), H256>>,
    pub code: RwLock<HashMap<Address, Bytes>>,
    pub execution_result: RwLock<Option<Result<Executed, CallError>>>,
    pub receipts: RwLock<HashMap<TransactionId, LocalizedReceipt>>,
    pub logs: RwLock<Vec<LocalizedLogEntry>>,
    pub queue_size: AtomicUsize,
    pub miner: Arc<Miner>,
    pub spec: Spec,
    pub vm_factory: EvmFactory,
    pub latest_block_timestamp: RwLock<u64>,
    pub ancient_block: RwLock<Option<(H256, u64)>>,
    pub first_block: RwLock<Option<(H256, u64)>>,
    pub traces: RwLock<Option<Vec<LocalizedTrace>>>,
    pub history: RwLock<Option<u64>>,
}

Test client.

Fields
blocks: RwLock<HashMap<H256, Bytes>>

Blocks.

numbers: RwLock<HashMap<usize, H256>>

Mapping of numbers to hashes.

genesis_hash: H256

Genesis block hash.

last_hash: RwLock<H256>

Last block hash.

extra_data: Bytes

Extra data do set for each block

difficulty: RwLock<U256>

Difficulty.

balances: RwLock<HashMap<Address, U256>>

Balances.

nonces: RwLock<HashMap<Address, U256>>

Nonces.

storage: RwLock<HashMap<(Address, H256), H256>>

Storage.

code: RwLock<HashMap<Address, Bytes>>

Code.

execution_result: RwLock<Option<Result<Executed, CallError>>>

Execution result.

receipts: RwLock<HashMap<TransactionId, LocalizedReceipt>>

Transaction receipts.

logs: RwLock<Vec<LocalizedLogEntry>>

Logs

queue_size: AtomicUsize

Block queue size.

miner: Arc<Miner>

Miner

spec: Spec

Spec

vm_factory: EvmFactory

VM Factory

latest_block_timestamp: RwLock<u64>

Timestamp assigned to latest sealed block

ancient_block: RwLock<Option<(H256, u64)>>

Ancient block info.

first_block: RwLock<Option<(H256, u64)>>

First block info.

traces: RwLock<Option<Vec<LocalizedTrace>>>

Traces to return

history: RwLock<Option<u64>>

Pruning history size to report.

Methods

impl TestBlockChainClient

fn new() -> Self

Creates new test client.

fn new_with_extra_data(extra_data: Bytes) -> Self

Creates new test client with specified extra data for each block

fn new_with_spec(spec: Spec) -> Self

Create test client with custom spec.

fn new_with_spec_and_extra(spec: Spec, extra_data: Bytes) -> Self

Create test client with custom spec and extra data.

fn set_transaction_receipt(&self, id: TransactionId, receipt: LocalizedReceipt)

Set the transaction receipt result

fn set_execution_result(&self, result: Result<Executed, CallError>)

Set the execution result.

fn set_balance(&self, address: Address, balance: U256)

Set the balance of account address to balance.

fn set_nonce(&self, address: Address, nonce: U256)

Set nonce of account address to nonce.

fn set_code(&self, address: Address, code: Bytes)

Set code at address.

fn set_storage(&self, address: Address, position: H256, value: H256)

Set storage position to value for account address.

fn set_queue_size(&self, size: usize)

Set block queue size for testing

fn set_latest_block_timestamp(&self, ts: u64)

Set timestamp assigned to latest sealed block

fn set_logs(&self, logs: Vec<LocalizedLogEntry>)

Set logs to return for each logs call.

fn add_blocks(&self, count: usize, with: EachBlockWith)

Add blocks to test client.

fn corrupt_block(&self, n: BlockNumber)

Make a bad block by setting invalid extra data.

fn corrupt_block_parent(&self, n: BlockNumber)

Make a bad block by setting invalid parent hash.

fn block_hash_delta_minus(&mut self, delta: usize) -> H256

TODO:

fn insert_transaction_to_queue(&self)

Inserts a transaction to miners transactions queue.

fn set_history(&self, h: Option<u64>)

Set reported history size.