Trait ethcore::client::BlockChainClient

pub trait BlockChainClient: Sync + Send {
    fn block_header(&self, hash: &H256) -> Option<Bytes>;
    fn block_body(&self, hash: &H256) -> Option<Bytes>;
    fn block(&self, hash: &H256) -> Option<Bytes>;
    fn block_status(&self, hash: &H256) -> BlockStatus;
    fn block_header_at(&self, n: BlockNumber) -> Option<Bytes>;
    fn block_body_at(&self, n: BlockNumber) -> Option<Bytes>;
    fn block_at(&self, n: BlockNumber) -> Option<Bytes>;
    fn block_status_at(&self, n: BlockNumber) -> BlockStatus;
    fn tree_route(&self, from: &H256, to: &H256) -> Option<TreeRoute>;
    fn state_data(&self, hash: &H256) -> Option<Bytes>;
    fn block_receipts(&self, hash: &H256) -> Option<Bytes>;
    fn import_block(&mut self, bytes: Bytes) -> ImportResult;
    fn queue_status(&self) -> BlockQueueStatus;
    fn clear_queue(&mut self);
    fn chain_info(&self) -> BlockChainInfo;
}

Blockchain database client. Owns and manages a blockchain and a block queue.

Required Methods

fn block_header(&self, hash: &H256) -> Option<Bytes>

Get raw block header data by block header hash.

fn block_body(&self, hash: &H256) -> Option<Bytes>

Get raw block body data by block header hash. Block body is an RLP list of two items: uncles and transactions.

fn block(&self, hash: &H256) -> Option<Bytes>

Get raw block data by block header hash.

fn block_status(&self, hash: &H256) -> BlockStatus

Get block status by block header hash.

fn block_header_at(&self, n: BlockNumber) -> Option<Bytes>

Get raw block header data by block number.

fn block_body_at(&self, n: BlockNumber) -> Option<Bytes>

Get raw block body data by block number. Block body is an RLP list of two items: uncles and transactions.

fn block_at(&self, n: BlockNumber) -> Option<Bytes>

Get raw block data by block number.

fn block_status_at(&self, n: BlockNumber) -> BlockStatus

Get block status by block number.

fn tree_route(&self, from: &H256, to: &H256) -> Option<TreeRoute>

Get a tree route between from and to. See BlockChain::tree_route.

fn state_data(&self, hash: &H256) -> Option<Bytes>

Get latest state node

fn block_receipts(&self, hash: &H256) -> Option<Bytes>

Get raw block receipts data by block header hash.

fn import_block(&mut self, bytes: Bytes) -> ImportResult

Import a block into the blockchain.

fn queue_status(&self) -> BlockQueueStatus

Get block queue information.

fn clear_queue(&mut self)

Clear block queue and abort all import activity.

fn chain_info(&self) -> BlockChainInfo

Get blockchain information.