Trait ethcore::client::BlockChainClient

pub trait BlockChainClient: Sync + Send {
    fn block_header(&self, id: BlockId) -> Option<Bytes>;
    fn block_body(&self, id: BlockId) -> Option<Bytes>;
    fn block(&self, id: BlockId) -> Option<Bytes>;
    fn block_status(&self, id: BlockId) -> BlockStatus;
    fn block_total_difficulty(&self, id: BlockId) -> Option<U256>;
    fn code(&self, address: &Address) -> Option<Bytes>;
    fn transaction(&self, id: TransactionId) -> Option<LocalizedTransaction>;
    fn tree_route(&self, from: &H256, to: &H256) -> Option<TreeRoute>;
    fn state_data(&self, hash: &H256) -> Option<Bytes>;
    fn block_receipts(&self, hash: &H256) -> Option<Bytes>;
    fn import_block(&self, bytes: Bytes) -> ImportResult;
    fn queue_info(&self) -> BlockQueueInfo;
    fn clear_queue(&self);
    fn chain_info(&self) -> BlockChainInfo;

    fn best_block_header(&self) -> Bytes { ... }
}

Blockchain database client. Owns and manages a blockchain and a block queue.

Required Methods

fn block_header(&self, id: BlockId) -> Option<Bytes>

Get raw block header data by block id.

fn block_body(&self, id: BlockId) -> Option<Bytes>

Get raw block body data by block id. Block body is an RLP list of two items: uncles and transactions.

fn block(&self, id: BlockId) -> Option<Bytes>

Get raw block data by block header hash.

fn block_status(&self, id: BlockId) -> BlockStatus

Get block status by block header hash.

fn block_total_difficulty(&self, id: BlockId) -> Option<U256>

Get block total difficulty.

fn code(&self, address: &Address) -> Option<Bytes>

Get address code.

fn transaction(&self, id: TransactionId) -> Option<LocalizedTransaction>

Get transaction with given hash.

fn tree_route(&self, from: &H256, to: &H256) -> Option<TreeRoute>

Get a tree route between from and to. See BlockChain::tree_route.

fn state_data(&self, hash: &H256) -> Option<Bytes>

Get latest state node

fn block_receipts(&self, hash: &H256) -> Option<Bytes>

Get raw block receipts data by block header hash.

fn import_block(&self, bytes: Bytes) -> ImportResult

Import a block into the blockchain.

fn queue_info(&self) -> BlockQueueInfo

Get block queue information.

fn clear_queue(&self)

Clear block queue and abort all import activity.

fn chain_info(&self) -> BlockChainInfo

Get blockchain information.

Provided Methods

fn best_block_header(&self) -> Bytes

Get the best block header.