Trait ethcore::client::BlockChainClient

pub trait BlockChainClient: Sync + Send {
    fn block_header(&self, id: BlockId) -> Option<Bytes>;
    fn block_body(&self, id: BlockId) -> Option<Bytes>;
    fn block(&self, id: BlockId) -> Option<Bytes>;
    fn block_status(&self, id: BlockId) -> BlockStatus;
    fn block_total_difficulty(&self, id: BlockId) -> Option<U256>;
    fn nonce(&self, address: &Address) -> U256;
    fn block_hash(&self, id: BlockId) -> Option<H256>;
    fn code(&self, address: &Address) -> Option<Bytes>;
    fn balance(&self, address: &Address) -> U256;
    fn storage_at(&self, address: &Address, position: &H256) -> H256;
    fn transaction(&self, id: TransactionId) -> Option<LocalizedTransaction>;
    fn uncle(&self, id: UncleId) -> Option<Header>;
    fn transaction_receipt(&self, id: TransactionId) -> Option<LocalizedReceipt>;
    fn tree_route(&self, from: &H256, to: &H256) -> Option<TreeRoute>;
    fn state_data(&self, hash: &H256) -> Option<Bytes>;
    fn block_receipts(&self, hash: &H256) -> Option<Bytes>;
    fn import_block(&self, bytes: Bytes) -> ImportResult;
    fn queue_info(&self) -> BlockQueueInfo;
    fn clear_queue(&self);
    fn chain_info(&self) -> BlockChainInfo;
    fn blocks_with_bloom(&self, bloom: &H2048, from_block: BlockId, to_block: BlockId) -> Option<Vec<BlockNumber>>;
    fn logs(&self, filter: Filter) -> Vec<LocalizedLogEntry>;
    fn prepare_sealing(&self, author: Address, gas_floor_target: U256, extra_data: Bytes, transactions: Vec<SignedTransaction>) -> (Option<ClosedBlock>, HashSet<H256>);
    fn try_seal(&self, block: LockedBlock, seal: Vec<Bytes>) -> Result<SealedBlock, LockedBlock>;
    fn call(&self, t: &SignedTransaction) -> Result<Executed, Error>;
    fn engine(&self) -> &Engine;
    fn filter_traces(&self, filter: TraceFilter) -> Option<Vec<LocalizedTrace>>;
    fn trace(&self, trace: TraceId) -> Option<LocalizedTrace>;
    fn transaction_traces(&self, trace: TransactionId) -> Option<Vec<LocalizedTrace>>;
    fn block_traces(&self, trace: BlockId) -> Option<Vec<LocalizedTrace>>;
    fn last_hashes(&self) -> LastHashes;

    fn best_block_header(&self) -> Bytes { ... }
    fn generate_seal(&self, block: &ExecutedBlock, accounts: Option<&AccountProvider>) -> Option<Vec<Bytes>> { ... }
}

Blockchain database client. Owns and manages a blockchain and a block queue.

Required Methods

fn block_header(&self, id: BlockId) -> Option<Bytes>

Get raw block header data by block id.

fn block_body(&self, id: BlockId) -> Option<Bytes>

Get raw block body data by block id. Block body is an RLP list of two items: uncles and transactions.

fn block(&self, id: BlockId) -> Option<Bytes>

Get raw block data by block header hash.

fn block_status(&self, id: BlockId) -> BlockStatus

Get block status by block header hash.

fn block_total_difficulty(&self, id: BlockId) -> Option<U256>

Get block total difficulty.

fn nonce(&self, address: &Address) -> U256

Get address nonce.

fn block_hash(&self, id: BlockId) -> Option<H256>

Get block hash.

fn code(&self, address: &Address) -> Option<Bytes>

Get address code.

fn balance(&self, address: &Address) -> U256

Get address balance.

fn storage_at(&self, address: &Address, position: &H256) -> H256

Get value of the storage at given position.

fn transaction(&self, id: TransactionId) -> Option<LocalizedTransaction>

Get transaction with given hash.

fn uncle(&self, id: UncleId) -> Option<Header>

Get uncle with given id.

fn transaction_receipt(&self, id: TransactionId) -> Option<LocalizedReceipt>

Get transaction receipt with given hash.

fn tree_route(&self, from: &H256, to: &H256) -> Option<TreeRoute>

Get a tree route between from and to. See BlockChain::tree_route.

fn state_data(&self, hash: &H256) -> Option<Bytes>

Get latest state node

fn block_receipts(&self, hash: &H256) -> Option<Bytes>

Get raw block receipts data by block header hash.

fn import_block(&self, bytes: Bytes) -> ImportResult

Import a block into the blockchain.

fn queue_info(&self) -> BlockQueueInfo

Get block queue information.

fn clear_queue(&self)

Clear block queue and abort all import activity.

fn chain_info(&self) -> BlockChainInfo

Get blockchain information.

fn blocks_with_bloom(&self, bloom: &H2048, from_block: BlockId, to_block: BlockId) -> Option<Vec<BlockNumber>>

Returns numbers of blocks containing given bloom.

fn logs(&self, filter: Filter) -> Vec<LocalizedLogEntry>

Returns logs matching given filter.

fn prepare_sealing(&self, author: Address, gas_floor_target: U256, extra_data: Bytes, transactions: Vec<SignedTransaction>) -> (Option<ClosedBlock>, HashSet<H256>)

Returns ClosedBlock prepared for sealing.

fn try_seal(&self, block: LockedBlock, seal: Vec<Bytes>) -> Result<SealedBlock, LockedBlock>

Attempts to seal given block. Returns SealedBlock on success and the same block in case of error.

fn call(&self, t: &SignedTransaction) -> Result<Executed, Error>

Makes a non-persistent transaction call.

fn engine(&self) -> &Engine

Executes a function providing it with a reference to an engine.

fn filter_traces(&self, filter: TraceFilter) -> Option<Vec<LocalizedTrace>>

Returns traces matching given filter.

fn trace(&self, trace: TraceId) -> Option<LocalizedTrace>

Returns trace with given id.

fn transaction_traces(&self, trace: TransactionId) -> Option<Vec<LocalizedTrace>>

Returns traces created by transaction.

fn block_traces(&self, trace: BlockId) -> Option<Vec<LocalizedTrace>>

Returns traces created by transaction from block.

fn last_hashes(&self) -> LastHashes

Get last hashes starting from best block.

Provided Methods

fn best_block_header(&self) -> Bytes

Get the best block header.

fn generate_seal(&self, block: &ExecutedBlock, accounts: Option<&AccountProvider>) -> Option<Vec<Bytes>>

Attempt to seal the block internally. See Engine.