Trait ethcore::client::ProvingBlockChainClient

pub trait ProvingBlockChainClient: BlockChainClient {
    fn prove_storage(&self,
                 key1: H256,
                 key2: H256,
                 from_level: u32,
                 id: BlockId)
                 -> Vec<Bytes>;
    fn prove_account(&self, key1: H256, from_level: u32, id: BlockId) -> Vec<Bytes>;
    fn code_by_hash(&self, account_key: H256, id: BlockId) -> Bytes;
}

Extended client interface for providing proofs of the state.

Required Methods

fn prove_storage(&self,
key1: H256,
key2: H256,
from_level: u32,
id: BlockId)
-> Vec<Bytes>

Prove account storage at a specific block id.

Both provided keys assume a secure trie. Returns a vector of raw trie nodes (in order from the root) proving the storage query. Nodes after from_level may be omitted. An empty vector indicates unservable query.

fn prove_account(&self, key1: H256, from_level: u32, id: BlockId) -> Vec<Bytes>

Prove account existence at a specific block id. The key is the keccak hash of the account's address. Returns a vector of raw trie nodes (in order from the root) proving the query. Nodes after from_level may be omitted. An empty vector indicates unservable query.

fn code_by_hash(&self, account_key: H256, id: BlockId) -> Bytes

Get code by address hash.