Struct ethcore::encoded::Header

pub struct Header(_);

Owning header view.

Methods

impl Header

fn new(encoded: Vec<u8>) -> Self

Create a new owning header view. Expects the data to be an RLP-encoded header -- any other case will likely lead to panics further down the line.

fn decode(&self) -> FullHeader

Upgrade this encoded view to a fully owned Header object.

fn view(&self) -> HeaderView

Get a borrowed header view onto the data.

fn rlp(&self) -> Rlp

Get the rlp of the header.

fn into_inner(self) -> Vec<u8>

Consume the view and return the raw bytes.

impl Header

fn hash(&self) -> H256

Returns the header hash.

fn parent_hash(&self) -> H256

Returns the parent hash.

fn uncles_hash(&self) -> H256

Returns the uncles hash.

fn author(&self) -> Address

Returns the author.

fn state_root(&self) -> H256

Returns the state root.

fn transactions_root(&self) -> H256

Returns the transaction trie root.

fn receipts_root(&self) -> H256

Returns the receipts trie root

fn log_bloom(&self) -> H2048

Returns the block log bloom

fn difficulty(&self) -> U256

Difficulty of this block

fn number(&self) -> BlockNumber

Number of this block.

fn timestamp(&self) -> u64

Time this block was produced.

fn gas_limit(&self) -> U256

Gas limit of this block.

fn gas_used(&self) -> U256

Total gas used in this block.

fn extra_data(&self) -> Vec<u8>

Block extra data.

fn seal(&self) -> Vec<Vec<u8>>

Engine-specific seal fields.