Enum ethcore::error::Error

pub enum Error {
    Client(ClientError),
    Util(UtilError),
    Block(BlockError),
    UnknownEngineName(String),
    Execution(ExecutionError),
    Transaction(TransactionError),
    Import(ImportError),
    PowHashInvalid,
    PowInvalid,
    Trie(TrieError),
    Io(IoError),
    StdIo(Error),
    Snappy(InvalidInput),
    Snapshot(SnapshotError),
    Ethkey(EthkeyError),
}

General error type which should be capable of representing all errors in ethcore.

Variants
Client(ClientError)

Client configuration error.

Util(UtilError)

Error concerning a utility.

Block(BlockError)

Error concerning block processing.

UnknownEngineName(String)

Unknown engine given.

Execution(ExecutionError)

Error concerning EVM code execution.

Transaction(TransactionError)

Error concerning transaction processing.

Import(ImportError)

Error concerning block import.

PowHashInvalid

PoW hash is invalid or out of date.

PowInvalid

The value of the nonce or mishash is invalid.

Trie(TrieError)

Error concerning TrieDBs

Io(IoError)

Io crate error.

StdIo(Error)

Standard io error.

Snappy(InvalidInput)

Snappy error.

Snapshot(SnapshotError)

Snapshot error.

Ethkey(EthkeyError)

Ethkey error.