Enum ethcore::error::ExecutionError

pub enum ExecutionError {
    NotEnoughBaseGas {
        required: U256,
        got: U256,
    },
    BlockGasLimitReached {
        gas_limit: U256,
        gas_used: U256,
        gas: U256,
    },
    InvalidNonce {
        expected: U256,
        got: U256,
    },
    NotEnoughCash {
        required: U512,
        got: U512,
    },
    Internal,
    TransactionMalformed(String),
}

Result of executing the transaction.

Variants
NotEnoughBaseGas

Returned when there gas paid for transaction execution is lower than base gas required.

Fields

required: U256

Absolute minimum gas required.

got: U256

Gas provided.

BlockGasLimitReached

Returned when block (gas_used + gas) > gas_limit.

If gas =< gas_limit, upstream may try to execute the transaction in next block.

Fields

gas_limit: U256

Gas limit of block for transaction.

gas_used: U256

Gas used in block prior to transaction.

gas: U256

Amount of gas in block.

InvalidNonce

Returned when transaction nonce does not match state nonce.

Fields

expected: U256

Nonce expected.

got: U256

Nonce found.

NotEnoughCash

Returned when cost of transaction (value + gas_price * gas) exceeds current sender balance.

Fields

required: U512

Minimum required balance.

got: U512

Actual balance.

Internal

Returned when internal evm error occurs.

TransactionMalformed(String)

Returned when generic transaction occurs