Enum ethcore::error::TransactionError

pub enum TransactionError {
    AlreadyImported,
    Old,
    TooCheapToReplace,
    LimitReached,
    InsufficientGasPrice {
        minimal: U256,
        got: U256,
    },
    InsufficientGas {
        minimal: U256,
        got: U256,
    },
    InsufficientBalance {
        balance: U256,
        cost: U256,
    },
    GasLimitExceeded {
        limit: U256,
        got: U256,
    },
    InvalidGasLimit(OutOfBounds<U256>),
    SenderBanned,
    RecipientBanned,
    CodeBanned,
    InvalidNetworkId,
}

Errors concerning transaction processing.

Variants
AlreadyImported

Transaction is already imported to the queue

Old

Transaction is not valid anymore (state already has higher nonce)

TooCheapToReplace

Transaction has too low fee (there is already a transaction with the same sender-nonce but higher gas price)

LimitReached

Transaction was not imported to the queue because limit has been reached.

InsufficientGasPrice

Transaction's gas price is below threshold.

Fields

minimal: U256

Minimal expected gas price

got: U256

Transaction gas price

InsufficientGas

Transaction's gas is below currently set minimal gas requirement.

Fields

minimal: U256

Minimal expected gas

got: U256

Transaction gas

InsufficientBalance

Sender doesn't have enough funds to pay for this transaction

Fields

balance: U256

Senders balance

cost: U256

Transaction cost

GasLimitExceeded

Transactions gas is higher then current gas limit

Fields

limit: U256

Current gas limit

got: U256

Declared transaction gas

InvalidGasLimit(OutOfBounds<U256>)

Transaction's gas limit (aka gas) is invalid.

SenderBanned

Transaction sender is banned.

RecipientBanned

Transaction receipient is banned.

CodeBanned

Contract creation code is banned.

InvalidNetworkId

Invalid network ID given.