Module ethcore::ethereum

Ethereum protocol module.

Contains all Ethereum network specific stuff, such as denominations and consensus specifications.

Reexports

pub use self::ethash::*;

pub use self::denominations::*;

Modules

denominations
TODO [Gav Wood] Please document me

ethash
TODO [Gav Wood] Please document me

Functions

new_frontier
Create a new Frontier mainnet chain spec.

new_frontier_like_test
Create a new Frontier main net chain spec without genesis accounts.

new_frontier_test
Create a new Frontier chain spec as though it never changes to Homestead.

new_homestead_test
Create a new Homestead chain spec as though it never changed from Frontier.

new_morden
Create a new Morden chain spec.

new_olympic
Create a new Olympic chain spec.