Module ethcore::ethereum

Ethereum protocol module.

Contains all Ethereum network specific stuff, such as denominations and consensus specifications.

Re-exports

pub use self::denominations::*;

pub use self::ethash::Ethash;

Modules

denominations

Export the denominations module.

ethash

Export the ethash module.

Functions

load

Load chain spec from SpecParams and JSON.

new_berlin_test

Create a new Foundation Berlin era spec.

new_berlin_to_london_test

Create a new BerlinToLondonAt5 era spec.

new_byzantium_test

Create a new Foundation Byzantium era spec.

new_byzantium_test_machine

Create a new Foundation Byzantium era spec.

new_byzantium_to_constantinoplefixat5_test

Create a new BizantiumToConstaninopleFixAt5 era spec.

new_callisto

Create a new Callisto chain spec

new_constantinople_fix_test

Create a new Foundation St. Peter’s (Contantinople Fix) era spec.

new_constantinople_fix_test_machine

Create a new Foundation St. Peter’s (Contantinople Fix) era spec.

new_constantinople_test

Create a new Foundation Constantinople era spec.

new_constantinople_test_machine

Create a new Foundation Constantinople era spec.

new_eip150_test

Create a new Foundation Homestead-EIP150-era chain spec as though it never changed from Homestead/Frontier.

new_eip161_test

Create a new Foundation Homestead-EIP161-era chain spec as though it never changed from Homestead/Frontier.

new_eip210_test_machine

Create a new Foundation Homestead-EIP210-era chain spec as though it never changed from Homestead/Frontier.

new_ellaism

Create a new Ellaism mainnet chain spec.

new_ewc

Create a new EWC mainnet chain spec.

new_foundation

Create a new Foundation mainnet chain spec.

new_frontier_test

Create a new Foundation Frontier-era chain spec as though it never changes to Homestead.

new_frontier_test_machine

Create a new Foundation Frontier-era chain spec as though it never changes to Homestead.

new_goerli

Create a new Görli testnet chain spec.

new_homestead_test

Create a new Foundation Homestead-era chain spec as though it never changed from Frontier.

new_homestead_test_machine

Create a new Foundation Homestead-era chain spec as though it never changed from Frontier.

new_istanbul_test

Create a new Foundation Istanbul era spec.

new_istanbul_test_machine

Create a new Foundation Istanbul era spec.

new_kovan

Create a new Kovan testnet chain spec.

new_kovan_wasm_test_machine

Create new Kovan spec with wasm activated at certain block

new_london_test

Create a new Foundation London era spec.

new_london_test_machine

Create a new Foundation London era chain spec.

new_mainnet_like

Create a new Foundation Mainnet chain spec without genesis accounts.

new_mcip3_test

Create a new Musicoin-MCIP3-era spec.

new_mcip3_test_machine

Create a new Musicoin-MCIP3-era spec.

new_mix

Create a new MIX mainnet chain spec.

new_morden

Create a new Morden testnet chain spec.

new_musicoin

Create a new Musicoin mainnet chain spec.

new_poanet

Create a new POA Network mainnet chain spec.

new_rinkeby

Create a new Rinkeby testnet chain spec.

new_ropsten

Create a new Ropsten testnet chain spec.

new_ropsten_test

Create a new Ropsten chain spec.

new_sokol

Create a new POA Sokol testnet chain spec.

new_transition_test

Create a new Foundation Frontier/Homestead/DAO chain spec with transition points at #5 and #8.

new_volta

Create a new Volta mainnet chain spec.

new_xdai

Create a new xDai chain spec.

new_yolo3

Create a new YOLO spec