Struct ethcore::ethereum::ethash::EthashParams

pub struct EthashParams {
    pub gas_limit_bound_divisor: U256,
    pub minimum_difficulty: U256,
    pub difficulty_bound_divisor: U256,
    pub difficulty_increment_divisor: u64,
    pub duration_limit: u64,
    pub block_reward: U256,
    pub registrar: Address,
    pub homestead_transition: u64,
    pub dao_hardfork_transition: u64,
    pub dao_hardfork_beneficiary: Address,
    pub dao_hardfork_accounts: Vec<Address>,
    pub difficulty_hardfork_transition: u64,
    pub difficulty_hardfork_bound_divisor: U256,
    pub bomb_defuse_transition: u64,
    pub eip150_transition: u64,
    pub eip155_transition: u64,
    pub eip160_transition: u64,
    pub eip161abc_transition: u64,
    pub eip161d_transition: u64,
    pub ecip1010_pause_transition: u64,
    pub ecip1010_continue_transition: u64,
    pub max_code_size: u64,
}

Ethash params.

Fields
gas_limit_bound_divisor: U256

Gas limit divisor.

minimum_difficulty: U256

Minimum difficulty.

difficulty_bound_divisor: U256

Difficulty bound divisor.

difficulty_increment_divisor: u64

Difficulty increment divisor.

duration_limit: u64

Block duration.

block_reward: U256

Block reward.

registrar: Address

Namereg contract address.

homestead_transition: u64

Homestead transition block number.

dao_hardfork_transition: u64

DAO hard-fork transition block (X).

dao_hardfork_beneficiary: Address

DAO hard-fork refund contract address (C).

dao_hardfork_accounts: Vec<Address>

DAO hard-fork DAO accounts list (L)

difficulty_hardfork_transition: u64

Transition block for a change of difficulty params (currently just bound_divisor).

difficulty_hardfork_bound_divisor: U256

Difficulty param after the difficulty transition.

bomb_defuse_transition: u64

Block on which there is no additional difficulty from the exponential bomb.

eip150_transition: u64

Number of first block where EIP-150 rules begin.

eip155_transition: u64

Number of first block where EIP-155 rules begin.

eip160_transition: u64

Number of first block where EIP-160 rules begin.

eip161abc_transition: u64

Number of first block where EIP-161.abc begin.

eip161d_transition: u64

Number of first block where EIP-161.d begins.

ecip1010_pause_transition: u64

Number of first block where ECIP-1010 begins.

ecip1010_continue_transition: u64

Number of first block where ECIP-1010 ends.

max_code_size: u64

Maximum amount of code that can be deploying into a contract.