Enum ethcore::executed::ExecutionError

pub enum ExecutionError {
    NotEnoughBaseGas {
        required: U256,
        got: U256,
    },
    BlockGasLimitReached {
        gas_limit: U256,
        gas_used: U256,
        gas: U256,
    },
    InvalidNonce {
        expected: U256,
        got: U256,
    },
    NotEnoughCash {
        required: U512,
        got: U512,
    },
    Internal,
    TransactionMalformed(String),
}

Result of executing the transaction.

Variants

NotEnoughBaseGas

Returned when there gas paid for transaction execution is lower than base gas required.

Fields

required

Absolute minimum gas required.

got

Gas provided.

BlockGasLimitReached

Returned when block (gas_used + gas) > gas_limit.

If gas =< gas_limit, upstream may try to execute the transaction in next block.

Fields

gas_limit

Gas limit of block for transaction.

gas_used

Gas used in block prior to transaction.

gas

Amount of gas in block.

InvalidNonce

Returned when transaction nonce does not match state nonce.

Fields

expected

Nonce expected.

got

Nonce found.

NotEnoughCash

Returned when cost of transaction (value + gas_price * gas) exceeds current sender balance.

Fields

required

Minimum required balance.

got

Actual balance.

Internal

Returned when internal evm error occurs.

TransactionMalformed

Returned when generic transaction occurs