Struct ethcore::header::Header

pub struct Header {
    pub parent_hash: H256,
    pub timestamp: u64,
    pub number: BlockNumber,
    pub author: Address,
    pub transactions_root: H256,
    pub uncles_hash: H256,
    pub extra_data: Bytes,
    pub state_root: H256,
    pub receipts_root: H256,
    pub log_bloom: LogBloom,
    pub gas_used: U256,
    pub gas_limit: U256,
    pub difficulty: U256,
    pub seal: Vec<Bytes>,
    pub hash: RefCell<Option<H256>>,
    pub bare_hash: RefCell<Option<H256>>,
}

A block header.

Reflects the specific RLP fields of a block in the chain with additional room for the seal which is non-specific.

Doesn't do all that much on its own.

Fields

parent_hash

TODO [Gav Wood] Please document me

timestamp

TODO [arkpar] Please document me

number

TODO [debris] Please document me

author

TODO [Gav Wood] Please document me

transactions_root

TODO [debris] Please document me

uncles_hash

TODO [debris] Please document me

extra_data

TODO [Gav Wood] Please document me

state_root

TODO [debris] Please document me

receipts_root

TODO [debris] Please document me

log_bloom

TODO [debris] Please document me

gas_used

TODO [debris] Please document me

gas_limit

TODO [Gav Wood] Please document me

difficulty

TODO [debris] Please document me

seal

TODO [arkpar] Please document me

hash

TODO [arkpar] Please document me

bare_hash

TODO [Gav Wood] Please document me

Methods

impl Header

fn new() -> Header

Create a new, default-valued, header.

fn number(&self) -> BlockNumber

TODO [Gav Wood] Please document me

fn timestamp(&self) -> u64

TODO [Gav Wood] Please document me

fn author(&self) -> &Address

TODO [Gav Wood] Please document me

fn extra_data(&self) -> &Bytes

TODO [Gav Wood] Please document me

fn state_root(&self) -> &H256

TODO [Gav Wood] Please document me

fn receipts_root(&self) -> &H256

TODO [Gav Wood] Please document me

fn gas_limit(&self) -> &U256

TODO [Gav Wood] Please document me

fn difficulty(&self) -> &U256

TODO [Gav Wood] Please document me

fn seal(&self) -> &Vec<Bytes>

TODO [Gav Wood] Please document me

fn set_number(&mut self, a: BlockNumber)

TODO [Gav Wood] Please document me

fn set_timestamp(&mut self, a: u64)

TODO [Gav Wood] Please document me

fn set_timestamp_now(&mut self)

TODO [Gav Wood] Please document me

fn set_author(&mut self, a: Address)

TODO [Gav Wood] Please document me

fn set_extra_data(&mut self, a: Bytes)

TODO [Gav Wood] Please document me

fn set_gas_used(&mut self, a: U256)

TODO [Gav Wood] Please document me

fn set_gas_limit(&mut self, a: U256)

TODO [Gav Wood] Please document me

fn set_difficulty(&mut self, a: U256)

TODO [Gav Wood] Please document me

fn set_seal(&mut self, a: Vec<Bytes>)

TODO [Gav Wood] Please document me

fn hash(&self) -> H256

Get the hash of this header (sha3 of the RLP).

fn bare_hash(&self) -> H256

Get the hash of the header excluding the seal

fn note_dirty(&self)

Note that some fields have changed. Resets the memoised hash.

fn stream_rlp(&self, s: &mut RlpStream, with_seal: Seal)

TODO [Gav Wood] Please document me

fn rlp(&self, with_seal: Seal) -> Bytes

TODO [Gav Wood] Please document me

fn rlp_sha3(&self, with_seal: Seal) -> H256

TODO [debris] Please document me