Struct ethcore::header::Header

pub struct Header { /* fields omitted */ }

A block header.

Reflects the specific RLP fields of a block in the chain with additional room for the seal which is non-specific.

Doesn't do all that much on its own.

Methods

impl Header

pub fn new() -> Self

Create a new, default-valued, header.

pub fn parent_hash(&self) -> &H256

Get the parent_hash field of the header.

pub fn timestamp(&self) -> u64

Get the timestamp field of the header.

pub fn number(&self) -> BlockNumber

Get the number field of the header.

pub fn author(&self) -> &Address

Get the author field of the header.

pub fn extra_data(&self) -> &Bytes

Get the extra data field of the header.

pub fn state_root(&self) -> &H256

Get the state root field of the header.

pub fn receipts_root(&self) -> &H256

Get the receipts root field of the header.

pub fn log_bloom(&self) -> &Bloom

Get the log bloom field of the header.

pub fn transactions_root(&self) -> &H256

Get the transactions root field of the header.

pub fn uncles_hash(&self) -> &H256

Get the uncles hash field of the header.

pub fn gas_used(&self) -> &U256

Get the gas used field of the header.

pub fn gas_limit(&self) -> &U256

Get the gas limit field of the header.

pub fn difficulty(&self) -> &U256

Get the difficulty field of the header.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn seal(&self) -> &[Bytes]

Get the seal field of the header.

pub fn decode_seal<'a, T: FromIterator<&'a [u8]>>(
&'a self
) -> Result<T, DecoderError>

Get the seal field with RLP-decoded values as bytes.

pub fn set_parent_hash(&mut self, a: H256)

Set the number field of the header.

pub fn set_uncles_hash(&mut self, a: H256)

Set the uncles hash field of the header.

pub fn set_state_root(&mut self, a: H256)

Set the state root field of the header.

pub fn set_transactions_root(&mut self, a: H256)

Set the transactions root field of the header.

pub fn set_receipts_root(&mut self, a: H256)

Set the receipts root field of the header.

pub fn set_log_bloom(&mut self, a: Bloom)

Set the log bloom field of the header.

pub fn set_timestamp(&mut self, a: u64)

Set the timestamp field of the header.

pub fn set_number(&mut self, a: BlockNumber)

Set the number field of the header.

pub fn set_author(&mut self, a: Address)

Set the author field of the header.

pub fn set_extra_data(&mut self, a: Bytes)

Set the extra data field of the header.

pub fn set_gas_used(&mut self, a: U256)

Set the gas used field of the header.

pub fn set_gas_limit(&mut self, a: U256)

Set the gas limit field of the header.

pub fn set_difficulty(&mut self, a: U256)

Set the difficulty field of the header.

pub fn set_seal(&mut self, a: Vec<Bytes>)

Set the seal field of the header.

pub fn compute_hash(&mut self) -> H256

Get & memoize the hash of this header (keccak of the RLP with seal).

pub fn hash(&self) -> H256

Get the hash of this header (keccak of the RLP with seal).

pub fn bare_hash(&self) -> H256

Get the hash of the header excluding the seal

pub fn encoded(&self) -> Header

Encode the header, getting a type-safe wrapper around the RLP.