Module ethcore::machine

Generalization of a state machine for a consensus engine.

Structs

AuxiliaryData

Auxiliary data fetcher for an Ethereum machine. In Ethereum-like machines there are two kinds of auxiliary data: bodies and receipts.

EthashExtensions

Ethash-specific extensions.

EthereumMachine

An ethereum-like state machine.

Enums

AuxiliaryRequest

Request for auxiliary data of a block.

Traits

Machine

Generalization of types surrounding blockchain-suitable state machines.

Type Definitions

Call

Type alias for a function we can make calls through synchronously. Returns the call result and state proof for each call.

ScheduleCreationRules

Special rules to be applied to the schedule.