Module ethcore::miner

Miner module Keeps track of transactions and mined block.

Usage example:

extern crate ethcore_util as util;
extern crate ethcore;
use std::env;
use util::network::{NetworkService, NetworkConfiguration};
use ethcore::client::{Client, ClientConfig};
use ethcore::miner::{Miner, MinerService};

fn main() {
    let miner: Miner = Miner::default();
    // get status
    assert_eq!(miner.status().transactions_in_pending_queue, 0);

    // Check block for sealing
    //assert!(miner.sealing_block(client.deref()).lock().unwrap().is_some());
}

Structs

AccountDetails
Details of account

ExternalMiner
External Miner.

Miner
Keeps track of transactions using priority queue and holds currently mined block.

MinerStatus
Mining status

TransactionQueue
TransactionQueue implementation

Enums

TransactionImportResult
Represents the result of importing transaction.

TransactionOrigin
Transaction origin

Traits

ExternalMinerService
External miner interface.

MinerService
Miner client API