Struct ethcore::miner::Miner

pub struct Miner { /* private fields */ }

Keeps track of transactions using priority queue and holds currently mined block. Handles preparing work for “work sealing” or seals “internally” if Engine does not require work.

Implementations

impl Miner

pub fn add_work_listener(&self, notifier: Box<dyn NotifyWork>)

Push listener that will handle new jobs

pub fn add_transactions_listener(&self, f: Box<dyn Fn(&[H256]) + Send + Sync>)

Set a callback to be notified about imported transactions’ hashes.

pub fn new<A: LocalAccounts + 'static>(
options: MinerOptions,
gas_pricer: GasPricer,
spec: &Spec,
accounts: A
) -> Self

Creates new instance of miner Arc.

pub fn new_for_tests(spec: &Spec, accounts: Option<HashSet<Address>>) -> Miner

Creates new instance of miner with given spec and accounts.

NOTE This should be only used for tests.

pub fn new_for_tests_force_sealing(
spec: &Spec,
accounts: Option<HashSet<Address>>,
force_sealing: bool
) -> Miner

Creates new instance of miner with given spec and accounts.

NOTE This should be only used for tests.

pub fn set_io_channel(&self, io_channel: IoChannel<ClientIoMessage>)

Sets IoChannel

pub fn set_in_chain_checker<C>(&self, chain: &Arc<C>) where
C: TransactionInfo + Send + Sync + 'static,

Sets in-blockchain checker for transactions.

pub fn clear(&self)

Clear all pending block states

pub fn update_transaction_queue_limits(
&self,
block_gas_limit: U256,
block_base_fee: Option<U256>,
allow_non_eoa_sender: bool
)

Updates transaction queue verification limits.

Limits consist of current block gas limit and minimal gas price.

pub fn service_transaction_checker(&self) -> Option<ServiceTransactionChecker>

Returns ServiceTransactionChecker