Struct ethcore::miner::Miner

pub struct Miner {
    // some fields omitted
}

Keeps track of transactions using priority queue and holds currently mined block. Handles preparing work for "work sealing" or seals "internally" if Engine does not require work.

Methods

impl Miner

fn with_spec(spec: &Spec) -> Miner

Creates new instance of miner without accounts, but with given spec.

fn new(options: MinerOptions, gas_pricer: GasPricer, spec: &Spec, accounts: Option<Arc<AccountProvider>>) -> Arc<Miner>

Creates new instance of a miner Arc.

fn clear(&self)

Clear all pending block states

fn pending_state(&self) -> Option<State>

Get Some clone() of the current pending block's state or None if we're not sealing.

fn pending_block(&self) -> Option<Block>

Get Some clone() of the current pending block's state or None if we're not sealing.