Struct ethcore::miner::MinerOptions

pub struct MinerOptions {
    pub force_sealing: bool,
    pub reseal_on_external_tx: bool,
    pub reseal_on_own_tx: bool,
    pub reseal_on_uncle: bool,
    pub reseal_min_period: Duration,
    pub reseal_max_period: Duration,
    pub pending_set: PendingSet,
    pub work_queue_size: usize,
    pub enable_resubmission: bool,
    pub infinite_pending_block: bool,
    pub tx_queue_strategy: PrioritizationStrategy,
    pub tx_queue_penalization: Penalization,
    pub tx_queue_no_unfamiliar_locals: bool,
    pub refuse_service_transactions: bool,
    pub pool_limits: Options,
    pub pool_verification_options: Options,
}

Configures the behaviour of the miner.

Fields
force_sealing: bool

Force the miner to reseal, even when nobody has asked for work.

reseal_on_external_tx: bool

Reseal on receipt of new external transactions.

reseal_on_own_tx: bool

Reseal on receipt of new local transactions.

reseal_on_uncle: bool

Reseal when new uncle block has been imported.

reseal_min_period: Duration

Minimum period between transaction-inspired reseals.

reseal_max_period: Duration

Maximum period between blocks (enables force sealing after that).

pending_set: PendingSet

Whether we should fallback to providing all the queue's transactions or just pending.

work_queue_size: usize

How many historical work packages can we store before running out?

enable_resubmission: bool

Can we submit two different solutions for the same block and expect both to result in an import?

infinite_pending_block: bool

Create a pending block with maximal possible gas limit. NOTE: Such block will contain all pending transactions but will be invalid if mined.

tx_queue_strategy: PrioritizationStrategy

Strategy to use for prioritizing transactions in the queue.

tx_queue_penalization: Penalization

Simple senders penalization.

tx_queue_no_unfamiliar_locals: bool

Do we want to mark transactions recieved locally (e.g. RPC) as local if we don't have the sending account?

refuse_service_transactions: bool

Do we refuse to accept service transactions even if sender is certified.

pool_limits: Options

Transaction pool limits.

pool_verification_options: Options

Initial transaction verification options.