Struct ethcore::miner::MinerOptions

pub struct MinerOptions {
    pub new_work_notify: Vec<String>,
    pub force_sealing: bool,
    pub reseal_on_external_tx: bool,
    pub reseal_on_own_tx: bool,
    pub reseal_min_period: Duration,
    pub tx_gas_limit: U256,
    pub tx_queue_size: usize,
    pub tx_queue_strategy: PrioritizationStrategy,
    pub pending_set: PendingSet,
    pub work_queue_size: usize,
    pub enable_resubmission: bool,
    pub tx_queue_gas_limit: GasLimit,
    pub tx_queue_banning: Banning,
}

Configures the behaviour of the miner.

Fields
new_work_notify: Vec<String>

URLs to notify when there is new work.

force_sealing: bool

Force the miner to reseal, even when nobody has asked for work.

reseal_on_external_tx: bool

Reseal on receipt of new external transactions.

reseal_on_own_tx: bool

Reseal on receipt of new local transactions.

reseal_min_period: Duration

Minimum period between transaction-inspired reseals.

tx_gas_limit: U256

Maximum amount of gas to bother considering for block insertion.

tx_queue_size: usize

Maximum size of the transaction queue.

tx_queue_strategy: PrioritizationStrategy

Strategy to use for prioritizing transactions in the queue.

pending_set: PendingSet

Whether we should fallback to providing all the queue's transactions or just pending.

work_queue_size: usize

How many historical work packages can we store before running out?

enable_resubmission: bool

Can we submit two different solutions for the same block and expect both to result in an import?

tx_queue_gas_limit: GasLimit

Global gas limit for all transaction in the queue except for local and retracted.

tx_queue_banning: Banning

Banning settings