Module ethcore::snapshot

Snapshot creation, restoration, and network service.

Documentation of the format can be found at https://github.com/ethcore/parity/wiki/%22PV64%22-Snapshot-Format

Reexports

pub use self::service::{Service, DatabaseRestore};

Modules

io
Snapshot i/o. Ways of writing and reading snapshots. This module supports writing and reading snapshots of two different formats: packed and loose. Packed snapshots are written to a single file, and loose snapshots are written to multiple files in one directory.

service
Snapshot network service implementation.

Structs

BlockRebuilder
Rebuilds the blockchain from chunks.

ManifestData
Manifest data.

Progress
A progress indicator for snapshots.

StateRebuilder
Used to rebuild the state trie piece by piece.

Watcher
A ChainNotify implementation which will trigger a snapshot event at certain block numbers.

Enums

Error
Snapshot-related errors.

RestorationStatus
Statuses for restorations.

Traits

SnapshotService
The interface for a snapshot network service. This handles: - restoration of snapshots to temporary databases. - responding to queries for snapshot manifests and chunks

Functions

chunk_blocks
Create and write out all block chunks to disk, returning a vector of all the hashes of block chunks created.

chunk_state
Walk the given state database starting from the given root, creating chunks and writing them out.

take_snapshot
Take a snapshot using the given blockchain, starting block hash, and database, writing into the given writer.

verify_old_block
Verify an old block with the given header, engine, blockchain, body. If always is set, it will perform the fullest verification possible. If not, it will take a random sample to determine whether it will do heavy or light verification.