Struct ethcore::snapshot::service::Service

pub struct Service { /* fields omitted */ }

SnapshotService implementation. This controls taking snapshots and restoring from them.

Methods

impl Service

fn new(params: ServiceParams) -> Result<Self, Error>

Create a new snapshot service from the given parameters.

fn reader(&self) -> RwLockReadGuard<Option<LooseReader>>

Get a reference to the snapshot reader.

fn tick(&self)

Tick the snapshot service. This will log any active snapshot being taken.

fn take_snapshot(&self, client: &Client, num: u64) -> Result<(), Error>

Take a snapshot at the block with the given number. calling this while a restoration is in progress or vice versa will lead to a race condition where the first one to finish will have their produced snapshot overwritten.

fn init_restore(&self,
manifest: ManifestData,
recover: bool)
-> Result<(), Error>

Initialize the restoration synchronously. The recover flag indicates whether to recover the restored snapshot.

fn feed_state_chunk(&self, hash: H256, chunk: &[u8])

Feed a state chunk to be processed synchronously.

fn feed_block_chunk(&self, hash: H256, chunk: &[u8])

Feed a block chunk to be processed synchronously.