Module ethcore::spec

TODO [debris] Please document me

Structs

GenesisAccount
Genesis account data. Does no thave a DB overlay cache

Spec
Parameters for a block chain; includes both those intrinsic to the design of the chain and those to be interpreted by the active chain engine.

Functions

gzip64res_to_json
Converts file from base64 gzipped bytes to json