Struct ethcore::spec::Spec

pub struct Spec {

    pub name: String,
    pub engine: Arc<dyn EthEngine>,
    pub data_dir: String,
    pub nodes: Vec<String>,
    pub parent_hash: H256,
    pub author: Address,
    pub difficulty: U256,
    pub gas_limit: U256,
    pub gas_used: U256,
    pub timestamp: u64,
    pub transactions_root: H256,
    pub receipts_root: H256,
    pub extra_data: Bytes,
    pub seal_rlp: Bytes,
    pub base_fee: Option<U256>,
    pub hard_forks: BTreeSet<BlockNumber>,
    /* private fields */

}

Parameters for a block chain; includes both those intrinsic to the design of the chain and those to be interpreted by the active chain engine.

Fields
name: String

User friendly spec name

engine: Arc<dyn EthEngine>

What engine are we using for this?

data_dir: String

Name of the subdir inside the main data dir to use for chain data and settings.

nodes: Vec<String>

Known nodes on the network in enode format.

parent_hash: H256

The genesis block’s parent hash field.

author: Address

The genesis block’s author field.

difficulty: U256

The genesis block’s difficulty field.

gas_limit: U256

The genesis block’s gas limit field.

gas_used: U256

The genesis block’s gas used field.

timestamp: u64

The genesis block’s timestamp field.

transactions_root: H256

Transactions root of the genesis block. Should be KECCAK_NULL_RLP.

receipts_root: H256

Receipts root of the genesis block. Should be KECCAK_NULL_RLP.

extra_data: Bytes

The genesis block’s extra data field.

seal_rlp: Bytes

Each seal field, expressed as RLP, concatenated.

base_fee: Option<U256>

Base fee,

hard_forks: BTreeSet<BlockNumber>

List of hard forks in the network.

Implementations

impl Spec

pub fn state_root(&self) -> H256

Return the state root for the genesis state, memoising accordingly.

pub fn params(&self) -> &CommonParams

Get common blockchain parameters.

pub fn nodes(&self) -> &[String]

Get the known knodes of the network in enode format.

pub fn network_id(&self) -> u64

Get the configured Network ID.

pub fn chain_id(&self) -> u64

Get the chain ID used for signing.

pub fn subprotocol_name(&self) -> String

Get the configured subprotocol name.

pub fn fork_block(&self) -> Option<(BlockNumber, H256)>

Get the configured network fork block.

pub fn genesis_header(&self) -> Header

Get the header of the genesis block.

pub fn genesis_block(&self) -> Bytes

Compose the genesis block for this chain.

pub fn overwrite_genesis_params(&mut self, g: Genesis)

Overwrite the genesis components.

pub fn set_genesis_state(&mut self, s: PodState) -> Result<(), Error>

Alter the value of the genesis state.

pub fn genesis_state(&self) -> &PodState

Return genesis state as Plain old data.

pub fn is_state_root_valid(&self) -> bool

Returns false if the memoized state root is invalid. true otherwise.

pub fn ensure_db_good<T: Backend>(
&self,
db: T,
factories: &Factories
) -> Result<T, Error>

Ensure that the given state DB has the trie nodes in for the genesis state.

pub fn load_machine<R: Read>(reader: R) -> Result<EthereumMachine, String>

Loads just the state machine from a json file.

pub fn load<'a, T: Into<SpecParams<'a>>, R>(
params: T,
reader: R
) -> Result<Self, String> where
R: Read,

Loads spec from json file. Provide factories for executing contracts and ensuring storage goes to the right place.

pub fn genesis_epoch_data(&self) -> Result<Vec<u8>, String>

initialize genesis epoch data, using in-memory database for constructor.

pub fn new_instant() -> Spec

Create a new Spec with InstantSeal consensus which does internal sealing (not requiring work).

pub fn new_test_round_rewrite_bytecode_transitions() -> Self