Struct ethcore::state::State

pub struct State {
    // some fields omitted
}

Representation of the entire state of all accounts in the system.

Methods

impl State

fn new(db: JournalDB, account_start_nonce: U256) -> State

Creates new state with empty state root

fn from_existing(db: JournalDB, root: H256, account_start_nonce: U256) -> State

Creates new state with existing state root

fn new_temp() -> State

Create temporary state object

fn drop(self) -> (H256, JournalDB)

Destroy the current object and return root and database.

fn root(&self) -> &H256

Return reference to root

fn db(&mut self) -> &mut JournalDB

Expose the underlying database; good to use for calling state.db().commit().

fn new_contract(&mut self, contract: &Address)

Create a new contract at address contract. If there is already an account at the address it will have its code reset, ready for init_code().

fn kill_account(&mut self, account: &Address)

Remove an existing account.

fn exists(&self, a: &Address) -> bool

Determine whether an account exists.

fn balance(&self, a: &Address) -> U256

Get the balance of account a.

fn nonce(&self, a: &Address) -> U256

Get the nonce of account a.

fn storage_at(&self, a: &Address, key: &H256) -> H256

Mutate storage of account a so that it is value for key.

fn code(&self, a: &Address) -> Option<Bytes>

Mutate storage of account a so that it is value for key.

fn add_balance(&mut self, a: &Address, incr: &U256)

Add incr to the balance of account a.

fn sub_balance(&mut self, a: &Address, decr: &U256)

Subtract decr from the balance of account a.

fn transfer_balance(&mut self, from: &Address, to: &Address, by: &U256)

Subtracts by from the balance of from and adds it to that of to.

fn inc_nonce(&mut self, a: &Address)

Increment the nonce of account a by 1.

fn set_storage(&mut self, a: &Address, key: H256, value: H256)

Mutate storage of account a so that it is value for key.

fn init_code(&mut self, a: &Address, code: Bytes)

Initialise the code of account a so that it is value for key. NOTE: Account should have been created with new_contract.

fn apply(&mut self, env_info: &EnvInfo, engine: &Engine, t: &Transaction) -> ApplyResult

Execute a given transaction. This will change the state accordingly.

fn revert(&mut self, backup: State)

TODO [debris] Please document me

fn as_json(&self) -> String

Convert into a JSON representation.

fn commit_into(db: &mut HashDB, root: &mut H256, accounts: &mut HashMap<Address, Option<Account>>)

Commit accounts to SecTrieDBMut. This is similar to cpp-ethereum's dev::eth::commit. accounts is mutable because we may need to commit the code or storage and record that.

fn commit(&mut self)

Commits our cached account changes into the trie.

fn populate_from(&mut self, accounts: PodState)

Populate the state from accounts.

fn to_hashmap_pod(&self) -> HashMap<Address, PodAccount>

Populate a PodAccount map from this state.

fn to_pod(&self) -> PodState

Populate a PodAccount map from this state.