Module ethcore::trace

Tracing

Modules

filter
Trace filters type definitions

flat
Flat trace module

localized
Localized traces type definitions

trace
Tracing datatypes.

Structs

AddressesFilter
Addresses filter.

Config
Traces config.

ExecutiveTracer
Simple executive tracer. Traces all calls and creates. Ignores delegatecalls.

ExecutiveVMTracer
Simple VM tracer. Traces all operations.

Filter
Traces filter.

FlatBlockTraces
Represents all traces produced by transactions in a single block.

FlatTrace
Trace localized in vector of traces produced by a single transaction.

FlatTransactionTraces
Represents all traces produced by a single transaction.

ImportRequest
Traces import request.

LocalizedTrace
Localized trace.

MemoryDiff
A diff of some chunk of memory.

NoopTracer
Nonoperative tracer. Does not trace anything.

NoopVMTracer
Nonoperative VM tracer. Does not trace anything.

StorageDiff
A diff of some storage value.

TraceDB
Trace database.

VMExecutedOperation
A record of an executed VM operation.

VMOperation
A record of the execution of a single VM operation.

VMTrace
A record of a full VM trace for a CALL/CREATE.

Enums

Error
TraceDB errors.

TraceError
Trace evm errors.

Traits

Database
Db provides an interface to query tracesdb.

DatabaseExtras
DbExtras provides an interface to query extra data which is not stored in tracesdb, but necessary to work correctly.

Tracer
This trait is used by executive to build traces.

VMTracer
Used by executive to build VM traces.