Struct ethcore::trace::filter::AddressesFilter

pub struct AddressesFilter { /* fields omitted */ }

Addresses filter.

Used to create bloom possibilities and match filters.

Methods

impl AddressesFilter

pub fn matches(&self, address: &Address) -> bool

Returns true if address matches one of the searched addresses.

pub fn matches_all(&self) -> bool

Returns true if this address filter matches everything.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn blooms(&self) -> Vec<Bloom>

Returns blooms of this addresses filter.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn with_blooms(&self, blooms: Vec<Bloom>) -> Vec<Bloom>

Returns vector of blooms zipped with blooms of this addresses filter.