Struct ethcore::trace::flat::FlatTrace

pub struct FlatTrace {
    pub action: Action,
    pub result: Res,
    pub subtraces: usize,
    pub trace_address: Vec<usize>,
}

Trace localized in vector of traces produced by a single transaction.

Parent and children indexes refer to positions in this vector.

Fields

action

Type of action performed by a transaction.

result

Result of this action.

subtraces

Number of subtraces.

trace_address

Exact location of trace.

[index in root, index in first CALL, index in second CALL, ...]