Struct ethcore::transaction::Transaction

pub struct Transaction {
    pub nonce: U256,
    pub gas_price: U256,
    pub gas: U256,
    pub action: Action,
    pub value: U256,
    pub data: Bytes,
    pub v: u8,
    pub r: U256,
    pub s: U256,
    // some fields omitted
}

A set of information describing an externally-originating message call or contract creation operation.

Fields

nonce

TODO [debris] Please document me

gas_price

TODO [debris] Please document me

gas

TODO [debris] Please document me

action

TODO [debris] Please document me

value

TODO [debris] Please document me

data

TODO [Gav Wood] Please document me

v

TODO [Gav Wood] Please document me

r

TODO [Gav Wood] Please document me

s

TODO [debris] Please document me

Methods

impl Transaction

fn new() -> Self

TODO [Gav Wood] Please document me

fn new_call(to: Address, value: U256, data: Bytes, gas: U256, gas_price: U256, nonce: U256) -> Transaction

Create a new message-call transaction.

fn new_create(value: U256, data: Bytes, gas: U256, gas_price: U256, nonce: U256) -> Transaction

Create a new contract-creation transaction.

fn nonce(&self) -> &U256

Get the nonce of the transaction.

fn gas_price(&self) -> &U256

Get the gas price of the transaction.

fn gas(&self) -> &U256

Get the gas of the transaction.

fn action(&self) -> &Action

Get the action of the transaction (Create or Call).

fn value(&self) -> &U256

Get the value of the transaction.

fn data(&self) -> &Bytes

Get the data of the transaction.

fn rlp_append_opt(&self, s: &mut RlpStream, with_seal: Seal)

Append object into RLP stream, optionally with or without the signature.

fn rlp_bytes_opt(&self, with_seal: Seal) -> Bytes

Get the RLP serialisation of the object, optionally with or without the signature.

impl Transaction

fn hash(&self) -> H256

Get the hash of this header (sha3 of the RLP).

fn note_dirty(&self)

Note that some fields have changed. Resets the memoised hash.

fn standard_v(&self) -> u8

0 is v is 27, 1 if 28, and 4 otherwise.

fn signature(&self) -> Signature

Construct a signature object from the sig.

fn message_hash(&self) -> H256

The message hash of the transaction.

fn sender(&self) -> Result<Address, Error>

Returns transaction sender.

fn sign(&mut self, secret: &Secret)

Signs the transaction as coming from sender.

fn signed(self, secret: &Secret) -> Transaction

Signs the transaction as coming from sender.

fn gas_required_for(is_create: bool, data: &[u8], schedule: &Schedule) -> u64

Get the transaction cost in gas for the given params.

fn gas_required(&self, schedule: &Schedule) -> u64

Get the transaction cost in gas for this transaction.

fn check_low_s(&self) -> Result<(), Error>

Checks whether the signature has a low 's' value.

fn validate(self, schedule: &Schedule, require_low: bool) -> Result<Transaction, Error>

Do basic validation, checking for valid signature and minimum gas,