Struct ethcore::transaction::Transaction

pub struct Transaction {
    pub nonce: U256,
    pub gas_price: U256,
    pub gas: U256,
    pub action: Action,
    pub value: U256,
    pub data: Bytes,
}

A set of information describing an externally-originating message call or contract creation operation.

Fields
nonce: U256

Nonce.

gas_price: U256

Gas price.

gas: U256

Gas paid up front for transaction execution.

action: Action

Action, can be either call or contract create.

value: U256

Transfered value.

data: Bytes

Transaction data.

Methods

impl Transaction

fn rlp_append_unsigned_transaction(&self, s: &mut RlpStream)

Append object with a without signature into RLP stream

impl Transaction

fn hash(&self) -> H256

The message hash of the transaction.

fn sign(self, secret: &Secret) -> SignedTransaction

Signs the transaction as coming from sender.

fn with_signature(self, sig: Signature) -> SignedTransaction

Signs the transaction with signature.

fn fake_sign(self, from: Address) -> SignedTransaction

Specify the sender; this won't survive the serialize/deserialize process, but can be cloned.

fn gas_required_for(is_create: bool, data: &[u8], schedule: &Schedule) -> u64

Get the transaction cost in gas for the given params.

fn gas_required(&self, schedule: &Schedule) -> u64

Get the transaction cost in gas for this transaction.