Trait ethcore::verification::queue::kind::Kind

pub trait Kind: 'static + Sized + Send + Sync {
    type Input: Sized + Send + BlockLike + MallocSizeOf;
    type Unverified: Sized + Send + BlockLike + MallocSizeOf;
    type Verified: Sized + Send + BlockLike + MallocSizeOf;
    fn create(
        input: Self::Input, 
        engine: &dyn EthEngine, 
        check_seal: bool
    ) -> Result<Self::Unverified, (Self::Input, Error)>;

    fn verify(
        unverified: Self::Unverified, 
        engine: &dyn EthEngine, 
        check_seal: bool
    ) -> Result<Self::Verified, Error>;
}

Defines transitions between stages of verification.

It starts with a fallible transformation from an “input” into the unverified item. This consists of quick, simply done checks as well as extracting particular data.

Then, there is a verify function which performs more expensive checks and produces the verified output.

For correctness, the hashes produced by each stage of the pipeline should be consistent.

Associated Types

type Input: Sized + Send + BlockLike + MallocSizeOf

The first stage: completely unverified.

type Unverified: Sized + Send + BlockLike + MallocSizeOf

The second stage: partially verified.

type Verified: Sized + Send + BlockLike + MallocSizeOf

The third stage: completely verified.

Required methods

fn create(
input: Self::Input,
engine: &dyn EthEngine,
check_seal: bool
) -> Result<Self::Unverified, (Self::Input, Error)>

Attempt to create the Unverified item from the input.

fn verify(
unverified: Self::Unverified,
engine: &dyn EthEngine,
check_seal: bool
) -> Result<Self::Verified, Error>

Attempt to verify the Unverified item using the given engine.