Struct ethcore::views::BlockView

pub struct BlockView<'a> { /* fields omitted */ }

View onto block rlp.

Methods

impl<'a> BlockView<'a>

fn new(bytes: &'a [u8]) -> BlockView<'a>

Creates new view onto block from raw bytes.

fn new_from_rlp(rlp: Rlp<'a>) -> BlockView<'a>

Creates new view onto block from rlp.

fn hash(&self) -> H256

Block header hash.

fn rlp(&self) -> &Rlp<'a>

Return reference to underlaying rlp.

fn header(&self) -> Header

Create new Header object from header rlp.

fn header_rlp(&self) -> Rlp

Return header rlp.

fn header_view(&self) -> HeaderView<'a>

Create new header view obto block head rlp.

fn transactions(&self) -> Vec<SignedTransaction>

Return List of transactions in given block.

fn localized_transactions(&self) -> Vec<LocalizedTransaction>

Return List of transactions with additional localization info.

fn transactions_count(&self) -> usize

Return number of transactions in given block, without deserializing them.

fn transaction_views(&self) -> Vec<TransactionView<'a>>

Return List of transactions in given block.

fn transaction_hashes(&self) -> Vec<H256>

Return transaction hashes.

fn transaction_at(&self, index: usize) -> Option<SignedTransaction>

Returns transaction at given index without deserializing unnecessary data.

fn localized_transaction_at(&self, index: usize) -> Option<LocalizedTransaction>

Returns localized transaction at given index.

fn uncles(&self) -> Vec<Header>

Return list of uncles of given block.

fn uncles_count(&self) -> usize

Return number of uncles in given block, without deserializing them.

fn uncle_views(&self) -> Vec<HeaderView<'a>>

Return List of transactions in given block.

fn uncle_hashes(&self) -> Vec<H256>

Return list of uncle hashes of given block.

fn uncle_at(&self, index: usize) -> Option<Header>

Return nth uncle.

fn uncle_rlp_at(&self, index: usize) -> Option<Bytes>

Return nth uncle rlp.